chore: address issue #619
