+++ app/services/refraction_static_artifacts/__init__.py
@@ -5,6 +5,15 @@ from __future__ import annotations
 from app.services.refraction_static_artifacts import contract
 from app.services.refraction_static_artifacts import _legacy
 from app.services.refraction_static_artifacts._legacy import *  # noqa: F403
+from app.services.refraction_static_artifacts.cell_velocity import (
+    build_refraction_cell_solver_history_rows as build_refraction_cell_solver_history_rows,
+    build_refraction_refractor_velocity_grid_arrays as build_refraction_refractor_velocity_grid_arrays,
+    build_refraction_refractor_velocity_qc_payload as build_refraction_refractor_velocity_qc_payload,
+    write_refraction_cell_solver_history_csv as write_refraction_cell_solver_history_csv,
+    write_refraction_refractor_velocity_cells_csv as write_refraction_refractor_velocity_cells_csv,
+    write_refraction_refractor_velocity_grid_npz as write_refraction_refractor_velocity_grid_npz,
+    write_refraction_refractor_velocity_qc_json as write_refraction_refractor_velocity_qc_json,
+)
 from app.services.refraction_static_artifacts.grid_map import (
     build_refraction_grid_map_qc_arrays as build_refraction_grid_map_qc_arrays,
     build_refraction_grid_map_qc_payload as build_refraction_grid_map_qc_payload,
