+++ app/services/refraction_static_artifacts/_legacy.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 from collections.abc import Iterable, Mapping
-from dataclasses import replace
 import os
 from pathlib import Path
 from typing import Any
@@ -16,17 +15,8 @@ from app.services.refraction_static_cell_coordinates import (
     project_refraction_cell_points,
     refraction_cell_coordinate_metadata_from_config,
 )
-from app.services.refraction_static_cell_grid import build_refraction_cell_grid
-from app.services.refraction_static_cell_grid import assign_observation_midpoint_cells
-from app.services.refraction_static_design_matrix import (
-    LOW_FOLD_CELL_REJECTION_REASON,
-    LOW_FOLD_CELL_VELOCITY_STATUS,
-)
 from app.services.refraction_static_layer_config import (
     normalize_refraction_static_layers,
-)
-from app.services.refraction_static_layer_observations import (
-    build_refraction_layer_observation_masks_from_arrays,
 )
 from app.services.refraction_static_status import (
     REFRACTION_STATIC_STATUSES,
@@ -39,9 +29,7 @@ from app.services.refraction_static_t1lsst import (
     write_refraction_t1lsst_1layer_components_csv,
 )
 from app.services.refraction_static_types import (
-    RefractionLayerKind,
     RefractionDatumStaticsResult,
-    RefractionLayerSolveResult,
     RefractionStaticArtifactSet,
     ResolvedRefractionFirstLayer,
 )
@@ -54,7 +42,6 @@ from app.services.refraction_static_v1 import (
 )
 
 from app.services.refraction_static_artifacts.contract import (
-    _CELL_SOLVER_HISTORY_COLUMNS,
     _COMPONENT_COLUMNS,
     _FIRST_BREAK_FIT_QC_COLUMNS,
     _FIRST_BREAK_TIME_EXPORT_COLUMNS,
@@ -62,7 +49,6 @@ from app.services.refraction_static_artifacts.contract import (
     _NEAR_SURFACE_3LAYER_COLUMNS,
     _NEAR_SURFACE_COLUMNS,
     _REDUCED_TIME_QC_COLUMNS,
-    _REFRACTOR_VELOCITY_CELL_COLUMNS,
     _RESIDUAL_COLUMNS,
     _STATIC_COMPONENT_QC_ENDPOINT_COLUMNS,
     _STATIC_COMPONENT_QC_TRACE_COLUMNS,
@@ -133,7 +119,6 @@ from app.services.refraction_static_artifacts.formatters import (
     _csv_bool,
     _csv_cell_id,
     _csv_float,
-    _csv_grid_float,
     _csv_identifier,
     _csv_int,
     _csv_layer_index,
@@ -142,7 +127,6 @@ from app.services.refraction_static_artifacts.formatters import (
     _float_or_nan,
     _json_float,
     _nan_if_none,
-    _required_finite_float,
 )
 from app.services.refraction_static_artifacts.io import (
     _assert_strict_json,
@@ -158,14 +142,12 @@ from app.services.refraction_static_artifacts.stats import (
     _status_counts,
 )
 from app.services.refraction_static_artifacts.registry import (
-    _CELL_VELOCITY_COMPONENT_BY_LAYER,
     _artifact_entries_for_request,
     _artifact_list_for_qc,
     _build_manifest_payload,
     _cell_velocity_artifact_names,
     _cell_velocity_artifact_names_for_request as _cell_velocity_artifact_names_for_request,
     _cell_velocity_artifact_paths_for_request,
-    _cell_velocity_layer_kind,
     _request_cell_velocity_layer_kinds,
     _validate_declared_upstream_artifacts,
     _validate_upstream_artifact_names,
@@ -180,7 +162,6 @@ _refractor_cell_velocity_artifact_entries = (
 )
 _t1lsst_artifact_entries = _artifact_registry._t1lsst_artifact_entries
 _upstream_artifact_entries = _artifact_registry._upstream_artifact_entries
-
 
 def write_refraction_static_artifacts(
     *,
@@ -438,26 +419,30 @@ def write_refraction_static_artifacts(
         req=request,
         path=paths.static_history_json,
     )
+    from app.services.refraction_static_artifacts import (
+        cell_velocity as _cell_velocity_artifacts,
+    )
+
     for cell_artifacts in cell_velocity_artifact_paths:
-        write_refraction_refractor_velocity_cells_csv(
+        _cell_velocity_artifacts.write_refraction_refractor_velocity_cells_csv(
             result=values.result,
             req=request,
             path=cell_artifacts.cells_csv,
             layer_kind=cell_artifacts.layer_kind,
         )
-        write_refraction_refractor_velocity_grid_npz(
+        _cell_velocity_artifacts.write_refraction_refractor_velocity_grid_npz(
             result=values.result,
             req=request,
             path=cell_artifacts.grid_npz,
             layer_kind=cell_artifacts.layer_kind,
         )
-        write_refraction_refractor_velocity_qc_json(
+        _cell_velocity_artifacts.write_refraction_refractor_velocity_qc_json(
             result=values.result,
             req=request,
             path=cell_artifacts.qc_json,
             layer_kind=cell_artifacts.layer_kind,
         )
-        write_refraction_cell_solver_history_csv(
+        _cell_velocity_artifacts.write_refraction_cell_solver_history_csv(
             result=values.result,
             req=request,
             path=cell_artifacts.solver_history_csv,
@@ -549,7 +534,6 @@ def write_refraction_static_artifacts(
     _validate_declared_upstream_artifacts(root, upstream_names)
     return paths
 
-
 def write_refraction_static_solution_npz(
     *,
     result: RefractionDatumStaticsResult,
@@ -566,7 +550,6 @@ def write_refraction_static_solution_npz(
         resolved_first_layer=resolved_first_layer,
     )
     _write_npz_atomic(Path(path), payload)
-
 
 def write_refraction_static_qc_json(
     *,
@@ -592,7 +575,6 @@ def write_refraction_static_qc_json(
     _write_json_atomic(Path(path), payload)
     return payload
 
-
 def write_refraction_static_history_json(
     *,
     result: RefractionDatumStaticsResult,
@@ -611,7 +593,6 @@ def write_refraction_static_history_json(
     _write_json_atomic(Path(path), payload)
     return payload
 
-
 def write_refraction_statics_csv(
     *,
     result: RefractionDatumStaticsResult,
@@ -620,7 +601,6 @@ def write_refraction_statics_csv(
     values = _validate_result(result)
     rows = _trace_statics_rows(values.result)
     _write_csv_atomic(Path(path), _trace_statics_columns(values.result), rows)
-
 
 def write_near_surface_model_csv(
     *,
@@ -631,7 +611,6 @@ def write_near_surface_model_csv(
     rows = _near_surface_model_rows(values.result)
     _write_csv_atomic(Path(path), _near_surface_columns(values.result), rows)
 
-
 def write_first_break_residuals_csv(
     *,
     result: RefractionDatumStaticsResult,
@@ -641,7 +620,6 @@ def write_first_break_residuals_csv(
     values = _validate_result(result)
     rows = _first_break_residual_rows(values.result, req=req)
     _write_csv_atomic(Path(path), _RESIDUAL_COLUMNS, rows)
-
 
 def write_refraction_first_break_time_export_csv(
     *,
@@ -658,7 +636,6 @@ def write_refraction_first_break_time_export_csv(
     )
     _write_csv_atomic(Path(path), _FIRST_BREAK_TIME_EXPORT_COLUMNS, rows)
 
-
 def write_refraction_first_break_fit_qc_csv(
     *,
     result: RefractionDatumStaticsResult,
@@ -674,7 +651,6 @@ def write_refraction_first_break_fit_qc_csv(
     rows = _first_break_fit_qc_rows(arrays)
     _write_csv_atomic(Path(path), _FIRST_BREAK_FIT_QC_COLUMNS, rows)
 
-
 def write_refraction_first_break_fit_qc_npz(
     *,
     result: RefractionDatumStaticsResult,
@@ -688,7 +664,6 @@ def write_refraction_first_break_fit_qc_npz(
         req=request,
     )
     _write_npz_atomic(Path(path), arrays)
-
 
 def write_refraction_first_break_fit_qc_json(
     *,
@@ -705,7 +680,6 @@ def write_refraction_first_break_fit_qc_json(
     _write_json_atomic(Path(path), payload)
     return payload
 
-
 def write_refraction_reduced_time_qc_csv(
     *,
     result: RefractionDatumStaticsResult,
@@ -721,7 +695,6 @@ def write_refraction_reduced_time_qc_csv(
     rows = _reduced_time_qc_rows(arrays)
     _write_csv_atomic(Path(path), _REDUCED_TIME_QC_COLUMNS, rows)
 
-
 def write_refraction_reduced_time_qc_npz(
     *,
     result: RefractionDatumStaticsResult,
@@ -735,7 +708,6 @@ def write_refraction_reduced_time_qc_npz(
         req=request,
     )
     _write_npz_atomic(Path(path), arrays)
-
 
 def write_refraction_reduced_time_qc_json(
     *,
@@ -752,7 +724,6 @@ def write_refraction_reduced_time_qc_json(
     _write_json_atomic(Path(path), payload)
     return payload
 
-
 def write_refraction_static_components_csv(
     *,
     result: RefractionDatumStaticsResult,
@@ -761,7 +732,6 @@ def write_refraction_static_components_csv(
     values = _validate_result(result)
     rows = _component_rows(values.result)
     _write_csv_atomic(Path(path), _component_columns(values.result), rows)
-
 
 def write_refraction_static_component_qc_artifacts(
     *,
@@ -795,73 +765,22 @@ def write_refraction_static_component_qc_artifacts(
     _write_json_atomic(Path(json_path), payload)
     return payload
 
+def _cell_velocity_artifacts() -> Any:
+    from app.services.refraction_static_artifacts import cell_velocity
 
-def write_refraction_refractor_velocity_cells_csv(
-    *,
-    result: RefractionDatumStaticsResult,
-    req: RefractionStaticApplyRequest,
-    path: Path,
-    layer_kind: RefractionLayerKind | None = None,
-) -> None:
-    arrays = build_refraction_refractor_velocity_grid_arrays(
-        result=result,
-        req=req,
-        layer_kind=layer_kind,
-    )
-    rows = _refractor_velocity_cell_rows(arrays)
-    _write_csv_atomic(Path(path), _REFRACTOR_VELOCITY_CELL_COLUMNS, rows)
+    return cell_velocity
 
+def write_refraction_refractor_velocity_cells_csv(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), 'write_refraction_refractor_velocity_cells_csv')(*args, **kwargs)
 
-def write_refraction_refractor_velocity_grid_npz(
-    *,
-    result: RefractionDatumStaticsResult,
-    req: RefractionStaticApplyRequest,
-    path: Path,
-    layer_kind: RefractionLayerKind | None = None,
-) -> None:
-    arrays = build_refraction_refractor_velocity_grid_arrays(
-        result=result,
-        req=req,
-        layer_kind=layer_kind,
-    )
-    _validate_no_object_arrays(
-        arrays,
-        artifact_name=REFRACTION_REFRACTOR_VELOCITY_GRID_NPZ_NAME,
-    )
-    _write_npz_atomic(Path(path), arrays)
+def write_refraction_refractor_velocity_grid_npz(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), 'write_refraction_refractor_velocity_grid_npz')(*args, **kwargs)
 
+def write_refraction_refractor_velocity_qc_json(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), 'write_refraction_refractor_velocity_qc_json')(*args, **kwargs)
 
-def write_refraction_refractor_velocity_qc_json(
-    *,
-    result: RefractionDatumStaticsResult,
-    req: RefractionStaticApplyRequest,
-    path: Path,
-    layer_kind: RefractionLayerKind | None = None,
-) -> dict[str, Any]:
-    payload = build_refraction_refractor_velocity_qc_payload(
-        result=result,
-        req=req,
-        layer_kind=layer_kind,
-    )
-    _write_json_atomic(Path(path), payload)
-    return payload
-
-
-def write_refraction_cell_solver_history_csv(
-    *,
-    result: RefractionDatumStaticsResult,
-    req: RefractionStaticApplyRequest,
-    path: Path,
-    layer_kind: RefractionLayerKind | None = None,
-) -> None:
-    rows = build_refraction_cell_solver_history_rows(
-        result=result,
-        req=req,
-        layer_kind=layer_kind,
-    )
-    csv_rows = [_cell_solver_history_csv_row(row) for row in rows]
-    _write_csv_atomic(Path(path), _CELL_SOLVER_HISTORY_COLUMNS, csv_rows)
-
+def write_refraction_cell_solver_history_csv(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), 'write_refraction_cell_solver_history_csv')(*args, **kwargs)
 
 def write_refraction_grid_map_qc_csv(
     *,
@@ -875,7 +794,6 @@ def write_refraction_grid_map_qc_csv(
 
     _write_refraction_grid_map_qc_csv(result=result, req=req, path=path)
 
-
 def write_refraction_grid_map_qc_npz(
     *,
     result: RefractionDatumStaticsResult,
@@ -887,7 +805,6 @@ def write_refraction_grid_map_qc_npz(
     )
 
     _write_refraction_grid_map_qc_npz(result=result, req=req, path=path)
-
 
 def write_refraction_grid_map_qc_json(
     *,
@@ -901,7 +818,6 @@ def write_refraction_grid_map_qc_json(
 
     return _write_refraction_grid_map_qc_json(result=result, req=req, path=path)
 
-
 def build_refraction_grid_map_qc_arrays(
     *,
     result: RefractionDatumStaticsResult,
@@ -912,7 +828,6 @@ def build_refraction_grid_map_qc_arrays(
     )
 
     return _build_refraction_grid_map_qc_arrays(result=result, req=req)
-
 
 def build_refraction_grid_map_qc_payload(
     *,
@@ -925,551 +840,14 @@ def build_refraction_grid_map_qc_payload(
 
     return _build_refraction_grid_map_qc_payload(result=result, req=req)
 
+def build_refraction_refractor_velocity_grid_arrays(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), 'build_refraction_refractor_velocity_grid_arrays')(*args, **kwargs)
 
-def build_refraction_refractor_velocity_grid_arrays(
-    *,
-    result: RefractionDatumStaticsResult,
-    req: RefractionStaticApplyRequest,
-    layer_kind: RefractionLayerKind | None = None,
-) -> dict[str, np.ndarray]:
-    request = RefractionStaticApplyRequest.model_validate(req)
-    resolved_layer_kind = _cell_velocity_layer_kind(request, layer_kind=layer_kind)
-    result = _cell_velocity_artifact_result_for_layer(
-        result=result,
-        req=request,
-        layer_kind=resolved_layer_kind,
-    )
-    values = _validate_result(result)
-    if not _request_has_cell_velocity_layer(request):
-        raise RefractionStaticArtifactError(
-            'refractor cell velocity artifacts require solve_cell request mode'
-        )
-    if not _result_has_cell_velocity_arrays(values.result):
-        raise RefractionStaticArtifactError(
-            'refractor cell velocity artifacts require cell velocity result arrays'
-        )
-    refractor_cell = request.model.refractor_cell
-    if refractor_cell is None:
-        raise RefractionStaticArtifactError(
-            'model.refractor_cell is required for cell velocity artifacts'
-        )
-    component = _cell_velocity_component(resolved_layer_kind)
+def build_refraction_refractor_velocity_qc_payload(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), 'build_refraction_refractor_velocity_qc_payload')(*args, **kwargs)
 
-    grid_config = effective_refraction_cell_grid_config(refractor_cell)
-    grid = build_refraction_cell_grid(grid_config)
-    n_total_cells = int(grid.cell_id.shape[0])
-    active_cell_id = _required_cell_int_array(
-        values.result.active_cell_id,
-        name='active_cell_id',
-    )
-    inactive_cell_id = _required_cell_int_array(
-        values.result.inactive_cell_id,
-        name='inactive_cell_id',
-    )
-    cell_slowness = _required_cell_float_array(
-        values.result.cell_bedrock_slowness_s_per_m,
-        name='cell_bedrock_slowness_s_per_m',
-    )
-    cell_velocity = _required_cell_float_array(
-        values.result.cell_bedrock_velocity_m_s,
-        name='cell_bedrock_velocity_m_s',
-    )
-    cell_status = _required_cell_status_array(
-        values.result.cell_velocity_status,
-        name='cell_velocity_status',
-    )
-    row_midpoint_cell_id = _required_cell_int_array(
-        values.result.row_midpoint_cell_id,
-        name='row_midpoint_cell_id',
-    )
-    if row_midpoint_cell_id.shape != (values.n_rows,):
-        raise RefractionStaticArtifactError(
-            'row_midpoint_cell_id length must match residual rows'
-        )
-    cell_slowness = _layer_cell_values_for_active_cells(
-        cell_slowness,
-        active_cell_id=active_cell_id,
-        n_total_cells=n_total_cells,
-        name='cell_bedrock_slowness_s_per_m',
-    )
-    cell_velocity = _layer_cell_values_for_active_cells(
-        cell_velocity,
-        active_cell_id=active_cell_id,
-        n_total_cells=n_total_cells,
-        name='cell_bedrock_velocity_m_s',
-    )
-    cell_status = _layer_cell_status_for_active_cells(
-        cell_status,
-        active_cell_id=active_cell_id,
-        n_total_cells=n_total_cells,
-        name='cell_velocity_status',
-    )
-    active_shape = active_cell_id.shape
-    for name, array in (
-        ('cell_bedrock_slowness_s_per_m', cell_slowness),
-        ('cell_bedrock_velocity_m_s', cell_velocity),
-        ('cell_velocity_status', cell_status),
-    ):
-        if array.shape != active_shape:
-            raise RefractionStaticArtifactError(
-                f'{name} length must match active_cell_id'
-            )
-    _validate_refractor_velocity_cell_ids(
-        grid_cell_id=grid.cell_id,
-        active_cell_id=active_cell_id,
-        inactive_cell_id=inactive_cell_id,
-    )
-
-    active_cell_mask = np.zeros(n_total_cells, dtype=bool)
-    v2_m_s = np.full(n_total_cells, np.nan, dtype=np.float64)
-    slowness_s_per_m = np.full(n_total_cells, np.nan, dtype=np.float64)
-    velocity_status = np.full(n_total_cells, 'inactive', dtype='<U32')
-    smoothing_neighbor_count = _active_neighbor_count_by_cell(
-        active_cell_id=active_cell_id,
-        n_total_cells=n_total_cells,
-        number_of_cell_x=grid.number_of_cell_x,
-        number_of_cell_y=grid.number_of_cell_y,
-    )
-    for index, raw_cell_id in enumerate(active_cell_id.tolist()):
-        cell_id = int(raw_cell_id)
-        active_cell_mask[cell_id] = True
-        v2_m_s[cell_id] = float(cell_velocity[index])
-        slowness_s_per_m[cell_id] = float(cell_slowness[index])
-        velocity_status[cell_id] = str(cell_status[index])
-    low_fold_cell_id = _qc_cell_id_array(
-        values.result.qc,
-        'low_fold_cell_id',
-        n_total_cells=n_total_cells,
-    )
-    if low_fold_cell_id.size:
-        velocity_status[low_fold_cell_id] = LOW_FOLD_CELL_VELOCITY_STATUS
-
-    coordinate_mode = str(refractor_cell.coordinate_mode)
-    center_aliases = _refractor_cell_center_alias_arrays(
-        x_center_m=grid.x_center_m,
-        y_center_m=grid.y_center_m,
-        coordinate_mode=coordinate_mode,
-        line_origin_x_m=refractor_cell.line_origin_x_m,
-        line_origin_y_m=refractor_cell.line_origin_y_m,
-        line_azimuth_deg=refractor_cell.line_azimuth_deg,
-    )
-    initial_v2 = np.full(
-        n_total_cells,
-        _initial_cell_v2_m_s(request, layer_kind=resolved_layer_kind),
-        dtype=np.float64,
-    )
-    v2_update = np.full(n_total_cells, np.nan, dtype=np.float64)
-    finite_update = np.isfinite(v2_m_s) & np.isfinite(initial_v2)
-    v2_update[finite_update] = v2_m_s[finite_update] - initial_v2[finite_update]
-    smoothing_weight = _history_smoothing_weight(
-        request,
-        layer_kind=resolved_layer_kind,
-    )
-    smoothing_enabled = bool(smoothing_weight > 0.0)
-
-    n_observations = np.zeros(n_total_cells, dtype=np.int64)
-    n_used_observations = np.zeros(n_total_cells, dtype=np.int64)
-    cell_candidate_row = _cell_velocity_candidate_row_mask(
-        values.result,
-        request,
-        layer_kind=resolved_layer_kind,
-    )
-    cell_row_midpoint_cell_id = np.where(
-        cell_candidate_row,
-        row_midpoint_cell_id,
-        -1,
-    )
-    valid_row_cell = (
-        (cell_row_midpoint_cell_id >= 0)
-        & (cell_row_midpoint_cell_id < n_total_cells)
-    )
-    np.add.at(n_observations, cell_row_midpoint_cell_id[valid_row_cell], 1)
-    qc_observation_count = _qc_cell_count_array(
-        values.result.qc,
-        'cell_observation_count',
-        n_total_cells=n_total_cells,
-    )
-    if qc_observation_count is not None:
-        n_observations = qc_observation_count
-    used_row = valid_row_cell & np.asarray(values.result.used_row_mask, dtype=bool)
-    np.add.at(n_used_observations, cell_row_midpoint_cell_id[used_row], 1)
-    if np.any(n_used_observations > n_observations):
-        raise RefractionStaticArtifactError(
-            'used observations per cell cannot exceed total observations per cell'
-        )
-    n_sources = _unique_observation_endpoint_count_by_cell(
-        row_midpoint_cell_id=cell_row_midpoint_cell_id,
-        endpoint_id=values.result.row_source_node_id,
-        n_total_cells=n_total_cells,
-    )
-    n_receivers = _unique_observation_endpoint_count_by_cell(
-        row_midpoint_cell_id=cell_row_midpoint_cell_id,
-        endpoint_id=values.result.row_receiver_node_id,
-        n_total_cells=n_total_cells,
-    )
-    residual_stats = _per_cell_residual_stats_ms(
-        row_midpoint_cell_id=cell_row_midpoint_cell_id,
-        residual_time_s=values.result.residual_time_s,
-        used_row_mask=used_row,
-        n_total_cells=n_total_cells,
-    )
-    status_reason = _refractor_cell_status_reasons(
-        velocity_status=velocity_status,
-        n_observations=n_observations,
-    )
-
-    arrays = {
-        'cell_id': np.ascontiguousarray(grid.cell_id, dtype=np.int64),
-        'ix': np.ascontiguousarray(grid.ix, dtype=np.int64),
-        'iy': np.ascontiguousarray(grid.iy, dtype=np.int64),
-        'cell_ix': np.ascontiguousarray(grid.ix, dtype=np.int64),
-        'cell_iy': np.ascontiguousarray(grid.iy, dtype=np.int64),
-        'coordinate_mode': _string_array(
-            np.full(n_total_cells, coordinate_mode, dtype=f'<U{len(coordinate_mode)}')
-        ),
-        'x_min_m': np.ascontiguousarray(grid.x_min_m, dtype=np.float64),
-        'x_max_m': np.ascontiguousarray(grid.x_max_m, dtype=np.float64),
-        'y_min_m': np.ascontiguousarray(grid.y_min_m, dtype=np.float64),
-        'y_max_m': np.ascontiguousarray(grid.y_max_m, dtype=np.float64),
-        'x_center_m': np.ascontiguousarray(grid.x_center_m, dtype=np.float64),
-        'y_center_m': np.ascontiguousarray(grid.y_center_m, dtype=np.float64),
-        'cell_center_x_m': center_aliases['x_m'],
-        'cell_center_y_m': center_aliases['y_m'],
-        'cell_center_inline_m': center_aliases['inline_m'],
-        'cell_center_crossline_m': center_aliases['crossline_m'],
-        'active_cell_mask': np.ascontiguousarray(active_cell_mask, dtype=bool),
-        'n_observations_per_cell': np.ascontiguousarray(
-            n_observations,
-            dtype=np.int64,
-        ),
-        'n_used_observations_per_cell': np.ascontiguousarray(
-            n_used_observations,
-            dtype=np.int64,
-        ),
-        'n_rejected_observations_per_cell': np.ascontiguousarray(
-            n_observations - n_used_observations,
-            dtype=np.int64,
-        ),
-        'n_sources_per_cell': np.ascontiguousarray(n_sources, dtype=np.int64),
-        'n_receivers_per_cell': np.ascontiguousarray(n_receivers, dtype=np.int64),
-        'cell_velocity_layer_kind': _string_array(
-            np.full(
-                n_total_cells,
-                resolved_layer_kind,
-                dtype=f'<U{len(resolved_layer_kind)}',
-            )
-        ),
-        'cell_velocity_component': _string_array(
-            np.full(n_total_cells, component, dtype=f'<U{len(component)}')
-        ),
-        'velocity_m_s': np.ascontiguousarray(v2_m_s, dtype=np.float64),
-        'v2_m_s': np.ascontiguousarray(v2_m_s, dtype=np.float64),
-        'slowness_s_per_m': np.ascontiguousarray(slowness_s_per_m, dtype=np.float64),
-        'initial_velocity_m_s': np.ascontiguousarray(initial_v2, dtype=np.float64),
-        'initial_v2_m_s': np.ascontiguousarray(initial_v2, dtype=np.float64),
-        'velocity_update_from_initial_m_s': np.ascontiguousarray(
-            v2_update,
-            dtype=np.float64,
-        ),
-        'v2_update_from_initial_m_s': np.ascontiguousarray(
-            v2_update,
-            dtype=np.float64,
-        ),
-        'velocity_status': _string_array(velocity_status),
-        'status_reason': _string_array(status_reason),
-        'residual_rms_ms': residual_stats['rms'],
-        'residual_mad_ms': residual_stats['mad'],
-        'residual_mean_ms': residual_stats['mean'],
-        'residual_p95_abs_ms': residual_stats['p95_abs'],
-        'smoothing_enabled': np.full(n_total_cells, smoothing_enabled, dtype=bool),
-        'smoothing_weight': np.full(
-            n_total_cells,
-            smoothing_weight,
-            dtype=np.float64,
-        ),
-        'smoothing_neighbor_count': np.ascontiguousarray(
-            smoothing_neighbor_count,
-            dtype=np.int64,
-        ),
-    }
-    _validate_no_object_arrays(
-        arrays,
-        artifact_name=REFRACTION_REFRACTOR_VELOCITY_GRID_NPZ_NAME,
-    )
-    return arrays
-
-
-def build_refraction_refractor_velocity_qc_payload(
-    *,
-    result: RefractionDatumStaticsResult,
-    req: RefractionStaticApplyRequest,
-    layer_kind: RefractionLayerKind | None = None,
-) -> dict[str, Any]:
-    request = RefractionStaticApplyRequest.model_validate(req)
-    resolved_layer_kind = _cell_velocity_layer_kind(request, layer_kind=layer_kind)
-    result = _cell_velocity_artifact_result_for_layer(
-        result=result,
-        req=request,
-        layer_kind=resolved_layer_kind,
-    )
-    arrays = build_refraction_refractor_velocity_grid_arrays(
-        result=result,
-        req=request,
-        layer_kind=resolved_layer_kind,
-    )
-    refractor_cell = request.model.refractor_cell
-    if refractor_cell is None:
-        raise RefractionStaticArtifactError(
-            'model.refractor_cell is required for cell velocity QC'
-        )
-    grid_config = effective_refraction_cell_grid_config(refractor_cell)
-    active_mask = np.asarray(arrays['active_cell_mask'], dtype=bool)
-    velocity = np.asarray(arrays['velocity_m_s'], dtype=np.float64)
-    active_velocity = velocity[active_mask & np.isfinite(velocity)]
-    n_total = int(arrays['cell_id'].shape[0])
-    n_active = int(np.count_nonzero(active_mask))
-    n_observations_in_grid = int(np.sum(arrays['n_observations_per_cell']))
-    if request.model.method == 'multilayer_time_term':
-        n_valid_observations = int(
-            np.count_nonzero(
-                _cell_velocity_candidate_row_mask(
-                    result,
-                    request,
-                    layer_kind=resolved_layer_kind,
-                )
-            )
-        )
-    else:
-        n_valid_observations = int(
-            np.count_nonzero(result.valid_observation_mask_sorted)
-        )
-    default_outside_observations = max(0, n_valid_observations - n_observations_in_grid)
-    n_used = int(np.sum(arrays['n_used_observations_per_cell']))
-    n_low_fold = int(
-        np.count_nonzero(
-            np.asarray(arrays['velocity_status']).astype(str, copy=False)
-            == LOW_FOLD_CELL_VELOCITY_STATUS
-        )
-    )
-    n_low_fold_rejected = int(
-        np.sum(
-            np.asarray(arrays['n_rejected_observations_per_cell'], dtype=np.int64)[
-                np.asarray(arrays['velocity_status']).astype(str, copy=False)
-                == LOW_FOLD_CELL_VELOCITY_STATUS
-            ]
-        )
-    )
-    n_smoothing_rows = _qc_int(
-        result.qc,
-        'n_cell_smoothing_rows',
-        default=_estimated_cell_smoothing_rows(
-            active_cell_mask=active_mask,
-            number_of_cell_x=int(grid_config.number_of_cell_x),
-            number_of_cell_y=int(grid_config.number_of_cell_y),
-            velocity_smoothing_weight=float(
-                _history_smoothing_weight(
-                    request,
-                    layer_kind=resolved_layer_kind,
-                )
-            ),
-        ),
-    )
-    payload: dict[str, Any] = {
-        'artifact_version': ARTIFACT_VERSION,
-        'bedrock_velocity_mode': 'solve_cell',
-        'cell_velocity_layer_kind': resolved_layer_kind,
-        'cell_velocity_component': _cell_velocity_component(resolved_layer_kind),
-        'cell_assignment_mode': refractor_cell.assignment_mode,
-        **refraction_cell_coordinate_metadata_from_config(refractor_cell),
-        'outside_grid_policy': refractor_cell.outside_grid_policy,
-        'number_of_cell_x': int(grid_config.number_of_cell_x),
-        'number_of_cell_y': int(grid_config.number_of_cell_y),
-        'size_of_cell_x_m': float(grid_config.size_of_cell_x_m),
-        'size_of_cell_y_m': _json_float(grid_config.size_of_cell_y_m),
-        'n_total_cells': n_total,
-        'n_active_cells': n_active,
-        'n_inactive_cells': int(n_total - n_active),
-        'min_observations_per_cell': _qc_int(
-            result.qc,
-            'min_observations_per_cell',
-            default=_cell_velocity_min_observations_per_cell(
-                request,
-                layer_kind=resolved_layer_kind,
-            ),
-        ),
-        'n_low_fold_cells': _qc_int(
-            result.qc,
-            'n_low_fold_cells',
-            default=n_low_fold,
-        ),
-        'n_observations_outside_grid': _qc_int(
-            result.qc,
-            'n_observations_outside_grid',
-            default=default_outside_observations,
-        ),
-        'n_observations_rejected_by_low_fold_cell': _qc_int(
-            result.qc,
-            'n_observations_rejected_by_low_fold_cell',
-            default=n_low_fold_rejected,
-        ),
-        'low_fold_cell_rejection_reason': str(
-            result.qc.get(
-                'low_fold_cell_rejection_reason',
-                LOW_FOLD_CELL_REJECTION_REASON,
-            )
-        ),
-        'n_used_observations': n_used,
-        'velocity_min_m_s': _stat(active_velocity, 'min'),
-        'velocity_median_m_s': _stat(active_velocity, 'median'),
-        'velocity_max_m_s': _stat(active_velocity, 'max'),
-        'velocity_smoothing_weight': _history_smoothing_weight(
-            request,
-            layer_kind=resolved_layer_kind,
-        ),
-        'smoothing_reference_distance_m': _qc_optional_float(
-            result.qc,
-            'smoothing_reference_distance_m',
-            default=refractor_cell.smoothing_reference_distance_m,
-        ),
-        'n_cell_smoothing_rows': n_smoothing_rows,
-    }
-    _assert_strict_json(
-        payload,
-        artifact_name=_cell_velocity_artifact_names(resolved_layer_kind).qc_json,
-    )
-    return payload
-
-
-def build_refraction_cell_solver_history_rows(
-    *,
-    result: RefractionDatumStaticsResult,
-    req: RefractionStaticApplyRequest,
-    layer_kind: RefractionLayerKind | None = None,
-) -> list[RefractionCellSolverHistoryRow]:
-    request = RefractionStaticApplyRequest.model_validate(req)
-    resolved_layer_kind = _cell_velocity_layer_kind(request, layer_kind=layer_kind)
-    result = _cell_velocity_artifact_result_for_layer(
-        result=result,
-        req=request,
-        layer_kind=resolved_layer_kind,
-    )
-    values = _validate_result(result)
-    if not _request_has_cell_velocity_layer(request):
-        raise RefractionStaticArtifactError(
-            'cell solver history artifact requires solve_cell request mode'
-        )
-    if not _result_has_cell_velocity_arrays(values.result):
-        raise RefractionStaticArtifactError(
-            'cell solver history artifact requires cell velocity result arrays'
-        )
-
-    arrays = build_refraction_refractor_velocity_grid_arrays(
-        result=values.result,
-        req=request,
-        layer_kind=resolved_layer_kind,
-    )
-    cell_counts = _cell_solver_history_cell_counts(arrays)
-    initial_v2 = _initial_cell_v2_m_s(request, layer_kind=resolved_layer_kind)
-    final_velocity = np.asarray(arrays['velocity_m_s'], dtype=np.float64)
-    active_mask = np.asarray(arrays['active_cell_mask'], dtype=bool)
-    active_final_v2 = final_velocity[active_mask & np.isfinite(final_velocity)]
-    update = np.asarray(
-        arrays['velocity_update_from_initial_m_s'],
-        dtype=np.float64,
-    )
-    active_update = update[active_mask & np.isfinite(update)]
-
-    cell_candidate_row = _cell_velocity_candidate_row_mask(
-        values.result,
-        request,
-        layer_kind=resolved_layer_kind,
-    )
-    cell_used_row = cell_candidate_row & np.asarray(
-        values.result.used_row_mask,
-        dtype=bool,
-    )
-    n_candidate = int(np.count_nonzero(cell_candidate_row))
-    n_used = int(np.count_nonzero(cell_used_row))
-    n_robust_rejected = int(
-        np.count_nonzero(
-            cell_candidate_row
-            & np.asarray(values.result.rejected_by_robust_mask, dtype=bool)
-        )
-    )
-    smoothing_weight = _history_smoothing_weight(
-        request,
-        layer_kind=resolved_layer_kind,
-    )
-    damping_weight = float(request.solver.damping)
-    robust_threshold = float(request.solver.robust.threshold)
-    robust_iteration_count = _history_robust_iteration_count(
-        values.result,
-        request,
-    )
-    residual_stats = _history_residual_stats_ms(
-        values.result,
-        used_row_mask=cell_used_row,
-    )
-
-    return [
-        RefractionCellSolverHistoryRow(
-            iteration=0,
-            stage='initial',
-            n_candidate_observations=n_candidate,
-            n_used_observations=n_candidate,
-            n_rejected_observations=0,
-            n_active_cells=cell_counts['active'],
-            n_low_fold_cells=cell_counts['low_fold'],
-            n_empty_cells=cell_counts['empty'],
-            residual_rms_ms=None,
-            residual_mad_ms=None,
-            max_abs_residual_ms=None,
-            median_velocity_m_s=initial_v2,
-            median_v2_m_s=initial_v2,
-            min_velocity_m_s=initial_v2,
-            min_v2_m_s=initial_v2,
-            max_velocity_m_s=initial_v2,
-            max_v2_m_s=initial_v2,
-            max_abs_velocity_update_m_s=0.0,
-            max_abs_v2_update_m_s=0.0,
-            smoothing_weight=smoothing_weight,
-            damping_weight=damping_weight,
-            robust_threshold=robust_threshold,
-            converged=False,
-            convergence_reason='initial_state',
-        ),
-        RefractionCellSolverHistoryRow(
-            iteration=max(1, robust_iteration_count),
-            stage='final',
-            n_candidate_observations=n_candidate,
-            n_used_observations=n_used,
-            n_rejected_observations=n_robust_rejected,
-            n_active_cells=cell_counts['active'],
-            n_low_fold_cells=cell_counts['low_fold'],
-            n_empty_cells=cell_counts['empty'],
-            residual_rms_ms=residual_stats['rms'],
-            residual_mad_ms=residual_stats['mad'],
-            max_abs_residual_ms=residual_stats['max_abs'],
-            median_velocity_m_s=_stat(active_final_v2, 'median'),
-            median_v2_m_s=_stat(active_final_v2, 'median'),
-            min_velocity_m_s=_stat(active_final_v2, 'min'),
-            min_v2_m_s=_stat(active_final_v2, 'min'),
-            max_velocity_m_s=_stat(active_final_v2, 'max'),
-            max_v2_m_s=_stat(active_final_v2, 'max'),
-            max_abs_velocity_update_m_s=_history_max_abs(active_update),
-            max_abs_v2_update_m_s=_history_max_abs(active_update),
-            smoothing_weight=smoothing_weight,
-            damping_weight=damping_weight,
-            robust_threshold=robust_threshold,
-            converged=True,
-            convergence_reason=_history_convergence_reason(
-                robust_iteration_count=robust_iteration_count,
-                n_robust_rejected_observations=n_robust_rejected,
-                smoothing_weight=smoothing_weight,
-            ),
-        ),
-    ]
-
+def build_refraction_cell_solver_history_rows(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), 'build_refraction_cell_solver_history_rows')(*args, **kwargs)
 
 def build_refraction_static_component_qc_arrays(
     *,
@@ -1816,7 +1194,6 @@ def build_refraction_static_component_qc_arrays(
         'endpoint_static_status': endpoint_static_status,
     }
 
-
 def build_refraction_static_component_qc_payload(
     *,
     arrays: dict[str, np.ndarray],
@@ -1917,7 +1294,6 @@ def build_refraction_static_component_qc_payload(
         artifact_name=REFRACTION_STATIC_COMPONENT_QC_JSON_NAME,
     )
     return payload
-
 
 def build_refraction_static_solution_arrays(
     *,
@@ -2400,7 +1776,6 @@ def build_refraction_static_solution_arrays(
     _validate_no_object_arrays(arrays, artifact_name=REFRACTION_STATIC_SOLUTION_NPZ_NAME)
     return arrays
 
-
 def build_refraction_first_break_fit_qc_arrays(
     *,
     result: RefractionDatumStaticsResult,
@@ -2526,7 +1901,6 @@ def build_refraction_first_break_fit_qc_arrays(
     )
     return arrays
 
-
 def build_refraction_first_break_fit_qc_payload(
     *,
     result: RefractionDatumStaticsResult,
@@ -2596,7 +1970,6 @@ def build_refraction_first_break_fit_qc_payload(
     }
     _assert_strict_json(payload, artifact_name=REFRACTION_FIRST_BREAK_FIT_QC_JSON_NAME)
     return payload
-
 
 def build_refraction_reduced_time_qc_arrays(
     *,
@@ -2696,7 +2069,6 @@ def build_refraction_reduced_time_qc_arrays(
     )
     return arrays
 
-
 def build_refraction_reduced_time_qc_payload(
     *,
     result: RefractionDatumStaticsResult,
@@ -2742,7 +2114,6 @@ def build_refraction_reduced_time_qc_payload(
     }
     _assert_strict_json(payload, artifact_name=REFRACTION_REDUCED_TIME_QC_JSON_NAME)
     return payload
-
 
 def build_refraction_static_qc_payload(
     *,
@@ -3037,7 +2408,6 @@ def build_refraction_static_qc_payload(
     _assert_strict_json(payload, artifact_name=REFRACTION_STATIC_QC_JSON_NAME)
     return payload
 
-
 def _final_layer_qc_payload(raw_layers: object) -> dict[str, Any]:
     if not isinstance(raw_layers, dict):
         return {}
@@ -3045,7 +2415,6 @@ def _final_layer_qc_payload(raw_layers: object) -> dict[str, Any]:
     if isinstance(nested_layers, dict):
         return dict(nested_layers)
     return dict(raw_layers)
-
 
 def _sign_convention_qc_payload(
     req: RefractionStaticApplyRequest,
@@ -3057,7 +2426,6 @@ def _sign_convention_qc_payload(
         'positive_shift': POSITIVE_SHIFT_DESCRIPTION,
         'negative_shift': NEGATIVE_SHIFT_DESCRIPTION,
     }
-
 
 def _source_depth_field_correction_qc(
     result: RefractionDatumStaticsResult,
@@ -3083,7 +2451,6 @@ def _source_depth_field_correction_qc(
     payload.setdefault('warnings', [])
     return payload
 
-
 def _uphole_field_correction_qc(
     result: RefractionDatumStaticsResult,
     req: RefractionStaticApplyRequest,
@@ -3104,7 +2471,6 @@ def _uphole_field_correction_qc(
     payload.setdefault('uphole_mode', req.field_corrections.uphole.mode)
     payload.setdefault('component_name', 'uphole_shift_s')
     return payload
-
 
 def _manual_static_field_correction_qc(
     result: RefractionDatumStaticsResult,
@@ -3130,7 +2496,6 @@ def _manual_static_field_correction_qc(
     )
     payload.setdefault('component_name', 'manual_static_shift_s')
     return payload
-
 
 def _field_correction_composition_qc(
     result: RefractionDatumStaticsResult,
@@ -3171,7 +2536,6 @@ def _field_correction_composition_qc(
     payload.setdefault('sign_convention', SIGN_CONVENTION)
     return payload
 
-
 def _field_correction_component_requested(
     req: RefractionStaticApplyRequest,
 ) -> bool:
@@ -3180,7 +2544,6 @@ def _field_correction_component_requested(
         or req.field_corrections.uphole.mode != 'none'
         or req.field_corrections.manual_static.mode != 'none'
     )
-
 
 def build_refraction_static_history_payload(
     *,
@@ -3210,7 +2573,6 @@ def build_refraction_static_history_payload(
     _assert_strict_json(payload, artifact_name=REFRACTION_STATIC_HISTORY_JSON_NAME)
     return payload
 
-
 def refraction_static_trace_shift_component_names(
     req: RefractionStaticApplyRequest,
 ) -> tuple[str, ...]:
@@ -3219,7 +2581,6 @@ def refraction_static_trace_shift_component_names(
     if _field_components_applied_to_trace_shift(request):
         components.extend(_requested_field_component_names(request))
     return tuple(components)
-
 
 def refraction_static_double_application_qc(
     *,
@@ -3233,7 +2594,6 @@ def refraction_static_double_application_qc(
         requested_components=refraction_static_trace_shift_component_names(request),
         source_meta=source_meta,
     )
-
 
 def static_history_double_application_qc(
     *,
@@ -3283,7 +2643,6 @@ def static_history_double_application_qc(
         'warnings': warnings,
     }
 
-
 def _static_history_components(
     req: RefractionStaticApplyRequest,
 ) -> list[dict[str, object]]:
@@ -3321,18 +2680,15 @@ def _static_history_components(
         )
     return components
 
-
 def _refraction_history_component_name(req: RefractionStaticApplyRequest) -> str:
     if req.conversion.mode in {'t1lsst_1layer', 't1lsst_multilayer'}:
         return 'refraction_t1lsst'
     return 'refraction'
 
-
 def _static_history_cumulative_shift_field(req: RefractionStaticApplyRequest) -> str:
     if _field_components_applied_to_trace_shift(req):
         return 'final_trace_shift_s_sorted'
     return 'refraction_trace_shift_s_sorted'
-
 
 def _field_components_applied_to_trace_shift(
     req: RefractionStaticApplyRequest,
@@ -3342,7 +2698,6 @@ def _field_components_applied_to_trace_shift(
         and req.field_corrections.composition.enabled
         and req.field_corrections.composition.apply_to_trace_shift
     )
-
 
 def _requested_field_component_names(
     req: RefractionStaticApplyRequest,
@@ -3355,7 +2710,6 @@ def _requested_field_component_names(
     if req.field_corrections.manual_static.mode != 'none':
         components.append('manual_static')
     return tuple(components)
-
 
 def _static_history_qc_from_result(
     result: RefractionDatumStaticsResult,
@@ -3382,7 +2736,6 @@ def _static_history_qc_from_result(
         'warnings': [],
     }
 
-
 def _lineage_component_names(
     source_meta: Mapping[str, object],
 ) -> tuple[set[str], set[str]]:
@@ -3408,7 +2761,6 @@ def _lineage_component_names(
     if isinstance(history, Mapping):
         _collect_history_components(history, existing=existing)
     return existing, suspected
-
 
 def _collect_lineage_components(
     values: Mapping[str, object],
@@ -3442,7 +2794,6 @@ def _collect_lineage_components(
         else:
             suspected.update({'source_depth', 'uphole', 'manual_static'})
 
-
 def _collect_history_components(
     history: Mapping[str, object],
     *,
@@ -3460,14 +2811,12 @@ def _collect_history_components(
         if canonical:
             existing.add(canonical)
 
-
 def _string_items(value: object) -> tuple[str, ...]:
     if isinstance(value, str):
         return (value,)
     if not isinstance(value, (list, tuple)):
         return ()
     return tuple(item for item in value if isinstance(item, str))
-
 
 def _canonical_static_history_component(value: object) -> str | None:
     if not isinstance(value, str) or not value:
@@ -3491,7 +2840,6 @@ def _canonical_static_history_component(value: object) -> str | None:
     if name in {'manual_static', 'manual_static_shift_s'}:
         return 'manual_static'
     return None
-
 
 def _double_application_message(
     *,
@@ -3517,7 +2865,6 @@ def _double_application_message(
         f'double_application_policy={policy}'
     )
 
-
 def _validate_job_dir(job_dir: Path) -> Path:
     try:
         root = Path(job_dir)
@@ -3530,7 +2877,6 @@ def _validate_job_dir(job_dir: Path) -> Path:
     if not os.access(root, os.W_OK):
         raise RefractionStaticArtifactError('job directory is not writable')
     return root
-
 
 def _validate_result(result: RefractionDatumStaticsResult) -> _ValidatedResult:
     if not isinstance(result, RefractionDatumStaticsResult):
@@ -3690,7 +3036,6 @@ def _validate_result(result: RefractionDatumStaticsResult) -> _ValidatedResult:
         n_rows=n_rows,
     )
 
-
 def _validate_optional_arrays(
     *,
     result: RefractionDatumStaticsResult,
@@ -3711,7 +3056,6 @@ def _validate_optional_arrays(
             raise RefractionStaticArtifactError(
                 f'{label} array length mismatch for {name}'
             )
-
 
 def _validate_solve_cell_local_v2_arrays(
     *,
@@ -3786,7 +3130,6 @@ def _validate_solve_cell_local_v2_arrays(
             'solve_cell row_midpoint_cell_id length mismatch'
         )
 
-
 def _validate_resolved_first_layer(
     *,
     result: RefractionDatumStaticsResult,
@@ -3834,7 +3177,6 @@ def _validate_resolved_first_layer(
         )
     return resolved_first_layer
 
-
 def _velocities_close(left: float, right: float) -> bool:
     return bool(
         np.isclose(
@@ -3844,7 +3186,6 @@ def _velocities_close(left: float, right: float) -> bool:
             atol=1.0e-6,
         )
     )
-
 
 _TRACE_ARRAY_NAMES = (
     'valid_observation_mask_sorted',
@@ -4004,7 +3345,6 @@ _FIELD_TOTAL_VALID_STATUSES = frozenset(
     {'ok', _FIELD_DISABLED_STATUS, _FIELD_NOT_APPLICABLE_STATUS}
 )
 
-
 def _trace_statics_columns(
     result: RefractionDatumStaticsResult,
 ) -> tuple[str, ...]:
@@ -4022,7 +3362,6 @@ def _trace_statics_columns(
         'refraction_trace_shift_ms',
         ('final_trace_shift_ms', 'trace_field_static_status'),
     )
-
 
 def _trace_statics_rows(result: RefractionDatumStaticsResult) -> list[dict[str, object]]:
     rows: list[dict[str, object]] = []
@@ -4076,7 +3415,6 @@ def _trace_statics_rows(result: RefractionDatumStaticsResult) -> list[dict[str, 
         rows.append(row)
     return rows
 
-
 def _static_component_qc_trace_rows(
     arrays: dict[str, np.ndarray],
 ) -> list[dict[str, object]]:
@@ -4128,7 +3466,6 @@ def _static_component_qc_trace_rows(
             }
         )
     return rows
-
 
 def _static_component_qc_endpoint_rows(
     arrays: dict[str, np.ndarray],
@@ -4208,7 +3545,6 @@ def _static_component_qc_endpoint_rows(
         )
     return rows
 
-
 def _near_surface_model_rows(result: RefractionDatumStaticsResult) -> list[dict[str, object]]:
     rows: list[dict[str, object]] = []
     node_sh1_m = _node_sh1_weathering_thickness_m(result)
@@ -4263,7 +3599,6 @@ def _near_surface_model_rows(result: RefractionDatumStaticsResult) -> list[dict[
                 )
         rows.append(row)
     return rows
-
 
 def _first_break_residual_rows(
     result: RefractionDatumStaticsResult,
@@ -4330,7 +3665,6 @@ def _first_break_residual_rows(
             }
         )
     return rows
-
 
 def _first_break_time_export_rows(
     result: RefractionDatumStaticsResult,
@@ -4403,7 +3737,6 @@ def _first_break_time_export_rows(
         )
     return rows
 
-
 def _first_break_fit_qc_rows(
     arrays: Mapping[str, np.ndarray],
 ) -> list[dict[str, object]]:
@@ -4458,7 +3791,6 @@ def _first_break_fit_qc_rows(
         )
     return rows
 
-
 def _reduced_time_qc_rows(
     arrays: Mapping[str, np.ndarray],
 ) -> list[dict[str, object]]:
@@ -4502,7 +3834,6 @@ def _reduced_time_qc_rows(
         )
     return rows
 
-
 def _reduced_time_layer_gate_flags(
     req: RefractionStaticApplyRequest,
     offset_m: np.ndarray,
@@ -4533,7 +3864,6 @@ def _reduced_time_layer_gate_flags(
         )
     return {key: np.ascontiguousarray(value, dtype=bool) for key, value in flags.items()}
 
-
 def _offset_gate_mask(
     offset_m: np.ndarray,
     *,
@@ -4552,7 +3882,6 @@ def _offset_gate_mask(
         mask &= offset <= float(max_offset_m)
     return np.ascontiguousarray(mask, dtype=bool)
 
-
 def _reduced_time_layer_gate_kind(
     *,
     layer_kind_by_row: np.ndarray,
@@ -4565,7 +3894,6 @@ def _reduced_time_layer_gate_kind(
         out[mask] = kind
         empty &= ~mask
     return np.ascontiguousarray(out, dtype='<U32')
-
 
 def _reduced_time_reduction_velocity_by_row(
     *,
@@ -4595,7 +3923,6 @@ def _reduced_time_reduction_velocity_by_row(
         )
     return np.ascontiguousarray(velocity, dtype=np.float64)
 
-
 def _configured_reduction_velocity_by_row(
     *,
     req: RefractionStaticApplyRequest,
@@ -4607,7 +3934,6 @@ def _configured_reduction_velocity_by_row(
     for layer_kind, velocity in velocity_by_kind.items():
         out[kind == layer_kind] = velocity
     return np.ascontiguousarray(out, dtype=np.float64)
-
 
 def _configured_initial_velocity_by_layer(
     req: RefractionStaticApplyRequest,
@@ -4627,7 +3953,6 @@ def _configured_initial_velocity_by_layer(
             values[config.kind] = velocity_f
     return values
 
-
 def _configured_v1_velocity(req: RefractionStaticApplyRequest) -> float | None:
     first_layer = req.model.first_layer
     if first_layer is not None and first_layer.weathering_velocity_m_s is not None:
@@ -4635,7 +3960,6 @@ def _configured_v1_velocity(req: RefractionStaticApplyRequest) -> float | None:
     if req.model.weathering_velocity_m_s is not None:
         return float(req.model.weathering_velocity_m_s)
     return None
-
 
 def _reduced_time_status(
     *,
@@ -4658,7 +3982,6 @@ def _reduced_time_status(
     status[missing_offset] = 'missing_offset'
     status[missing_velocity] = 'missing_reduction_velocity'
     return np.ascontiguousarray(status, dtype='<U32')
-
 
 def _reduced_time_gate_qc(req: RefractionStaticApplyRequest) -> dict[str, Any]:
     payload: dict[str, Any] = {
@@ -4686,7 +4009,6 @@ def _reduced_time_gate_qc(req: RefractionStaticApplyRequest) -> dict[str, Any]:
         }
     return payload
 
-
 def _source_job_id(
     source_job_id: str | None,
     req: RefractionStaticApplyRequest | None,
@@ -4696,12 +4018,10 @@ def _source_job_id(
         raw = getattr(req, 'source_job_id', None)
     return '' if raw is None else str(raw)
 
-
 def _first_break_export_layer_context(
     result: RefractionDatumStaticsResult,
 ) -> tuple[np.ndarray, np.ndarray]:
     return _residual_row_layer_context(result)
-
 
 def _row_endpoint_id_context(
     result: RefractionDatumStaticsResult,
@@ -4732,7 +4052,6 @@ def _row_endpoint_id_context(
             out[index] = value
     return np.ascontiguousarray(out, dtype=object)
 
-
 def _row_endpoint_float_context(
     result: RefractionDatumStaticsResult,
     *,
@@ -4761,7 +4080,6 @@ def _row_endpoint_float_context(
         if value is not None:
             out[index] = value
     return np.ascontiguousarray(out, dtype=np.float64)
-
 
 def _first_break_row_time_terms(
     result: RefractionDatumStaticsResult,
@@ -4825,7 +4143,6 @@ def _first_break_row_time_terms(
         np.ascontiguousarray(receiver, dtype=np.float64),
     )
 
-
 def _endpoint_time_lookup(
     endpoint_key: np.ndarray,
     values: np.ndarray | None,
@@ -4841,7 +4158,6 @@ def _endpoint_time_lookup(
         )
     }
 
-
 def _first_break_moveout_time_s(
     *,
     modeled_pick_time_s: np.ndarray,
@@ -4856,14 +4172,12 @@ def _first_break_moveout_time_s(
     out[invalid] = np.nan
     return np.ascontiguousarray(out, dtype=np.float64)
 
-
 def _midpoint_coordinate(left: np.ndarray, right: np.ndarray) -> np.ndarray:
     left_arr = np.asarray(left, dtype=np.float64)
     right_arr = np.asarray(right, dtype=np.float64)
     out = 0.5 * (left_arr + right_arr)
     out[~np.isfinite(left_arr) | ~np.isfinite(right_arr)] = np.nan
     return np.ascontiguousarray(out, dtype=np.float64)
-
 
 def _first_break_fit_inline_crossline(
     *,
@@ -4897,7 +4211,6 @@ def _first_break_fit_inline_crossline(
         np.ascontiguousarray(crossline, dtype=np.float64),
     )
 
-
 def _residual_row_layer_context(
     result: RefractionDatumStaticsResult,
 ) -> tuple[np.ndarray, np.ndarray]:
@@ -4920,7 +4233,6 @@ def _residual_row_layer_context(
             )
     return np.ascontiguousarray(kind), np.ascontiguousarray(index)
 
-
 def _residual_row_string_context(
     result: RefractionDatumStaticsResult,
     field: str,
@@ -4936,7 +4248,6 @@ def _residual_row_string_context(
         )
     return np.ascontiguousarray(out, dtype=object)
 
-
 def _residual_row_velocity_context(
     result: RefractionDatumStaticsResult,
 ) -> np.ndarray:
@@ -4950,7 +4261,6 @@ def _residual_row_velocity_context(
             )
         return np.ascontiguousarray(out, dtype=np.float64)
     return _row_velocity_from_cell_or_scalar(result, n_rows)
-
 
 def _row_velocity_from_cell_or_scalar(
     result: RefractionDatumStaticsResult,
@@ -4976,7 +4286,6 @@ def _row_velocity_from_cell_or_scalar(
         valid = (cell_id >= 0) & (cell_id < int(velocity.size))
         out[valid] = velocity[cell_id[valid]]
     return np.ascontiguousarray(out, dtype=np.float64)
-
 
 def _residual_row_cell_context(
     result: RefractionDatumStaticsResult,
@@ -5026,7 +4335,6 @@ def _residual_row_cell_context(
         np.ascontiguousarray(cell_iy, dtype=np.int64),
     )
 
-
 def _residual_cell_x_count(
     *,
     result: RefractionDatumStaticsResult,
@@ -5046,7 +4354,6 @@ def _residual_cell_x_count(
         return None
     return _required_positive_qc_int(result.qc, 'number_of_cell_x')
 
-
 def _residual_rejection_reason(
     *,
     used: bool,
@@ -5061,7 +4368,6 @@ def _residual_rejection_reason(
     if not used:
         return 'not_used'
     return 'ok'
-
 
 def _component_rows(result: RefractionDatumStaticsResult) -> list[dict[str, object]]:
     node_pick_count = _node_lookup(result.node_id, result.node_pick_count)
@@ -5203,7 +4509,6 @@ def _component_rows(result: RefractionDatumStaticsResult) -> list[dict[str, obje
         rows.append(row)
     return rows
 
-
 def _component_columns(result: RefractionDatumStaticsResult) -> tuple[str, ...]:
     columns = _COMPONENT_COLUMNS
     if _has_source_depth_field_correction(result):
@@ -5248,14 +4553,12 @@ def _component_columns(result: RefractionDatumStaticsResult) -> tuple[str, ...]:
         columns = _insert_after(columns, anchor, ('field_shift_ms', 'field_status'))
     return columns
 
-
 def _near_surface_columns(result: RefractionDatumStaticsResult) -> tuple[str, ...]:
     if _has_node_3layer_static_fields(result):
         return _NEAR_SURFACE_3LAYER_COLUMNS
     if _has_node_2layer_static_fields(result):
         return _NEAR_SURFACE_2LAYER_COLUMNS
     return _NEAR_SURFACE_COLUMNS
-
 
 def _has_source_depth_field_correction(
     result: RefractionDatumStaticsResult,
@@ -5273,7 +4576,6 @@ def _has_source_depth_field_correction(
         )
     return True
 
-
 def _has_uphole_field_correction(
     result: RefractionDatumStaticsResult,
 ) -> bool:
@@ -5289,7 +4591,6 @@ def _has_uphole_field_correction(
             'uphole field correction arrays must be provided together'
         )
     return True
-
 
 def _has_manual_static_field_correction(
     result: RefractionDatumStaticsResult,
@@ -5307,7 +4608,6 @@ def _has_manual_static_field_correction(
             'manual static field correction arrays must be provided together'
         )
     return True
-
 
 def _has_field_correction_composition(
     result: RefractionDatumStaticsResult,
@@ -5333,7 +4633,6 @@ def _has_field_correction_composition(
         )
     return True
 
-
 def _source_depth_m_array(result: RefractionDatumStaticsResult) -> np.ndarray:
     if not _has_source_depth_field_correction(result):
         return _disabled_field_float_array(int(result.source_endpoint_key.shape[0]))
@@ -5341,7 +4640,6 @@ def _source_depth_m_array(result: RefractionDatumStaticsResult) -> np.ndarray:
         result.source_depth_m,
         int(result.source_endpoint_key.shape[0]),
     )
-
 
 def _source_depth_shift_s_array(result: RefractionDatumStaticsResult) -> np.ndarray:
     if not _has_source_depth_field_correction(result):
@@ -5351,7 +4649,6 @@ def _source_depth_shift_s_array(result: RefractionDatumStaticsResult) -> np.ndar
         int(result.source_endpoint_key.shape[0]),
     )
 
-
 def _source_depth_status_array(result: RefractionDatumStaticsResult) -> np.ndarray:
     if not _has_source_depth_field_correction(result):
         return _disabled_field_status_array(int(result.source_endpoint_key.shape[0]))
@@ -5359,7 +4656,6 @@ def _source_depth_status_array(result: RefractionDatumStaticsResult) -> np.ndarr
         result.source_depth_status,
         int(result.source_endpoint_key.shape[0]),
     )
-
 
 def _source_uphole_time_s_array(result: RefractionDatumStaticsResult) -> np.ndarray:
     if not _has_uphole_field_correction(result):
@@ -5369,7 +4665,6 @@ def _source_uphole_time_s_array(result: RefractionDatumStaticsResult) -> np.ndar
         int(result.source_endpoint_key.shape[0]),
     )
 
-
 def _source_uphole_shift_s_array(result: RefractionDatumStaticsResult) -> np.ndarray:
     if not _has_uphole_field_correction(result):
         return _disabled_field_float_array(int(result.source_endpoint_key.shape[0]))
@@ -5378,7 +4673,6 @@ def _source_uphole_shift_s_array(result: RefractionDatumStaticsResult) -> np.nda
         int(result.source_endpoint_key.shape[0]),
     )
 
-
 def _source_uphole_status_array(result: RefractionDatumStaticsResult) -> np.ndarray:
     if not _has_uphole_field_correction(result):
         return _disabled_field_status_array(int(result.source_endpoint_key.shape[0]))
@@ -5386,7 +4680,6 @@ def _source_uphole_status_array(result: RefractionDatumStaticsResult) -> np.ndar
         result.source_uphole_status,
         int(result.source_endpoint_key.shape[0]),
     )
-
 
 def _source_manual_static_shift_s_array(
     result: RefractionDatumStaticsResult,
@@ -5398,7 +4691,6 @@ def _source_manual_static_shift_s_array(
         int(result.source_endpoint_key.shape[0]),
     )
 
-
 def _source_manual_static_status_array(
     result: RefractionDatumStaticsResult,
 ) -> np.ndarray:
@@ -5408,7 +4700,6 @@ def _source_manual_static_status_array(
         result.source_manual_static_status,
         int(result.source_endpoint_key.shape[0]),
     )
-
 
 def _receiver_manual_static_shift_s_array(
     result: RefractionDatumStaticsResult,
@@ -5420,7 +4711,6 @@ def _receiver_manual_static_shift_s_array(
         int(result.receiver_endpoint_key.shape[0]),
     )
 
-
 def _receiver_manual_static_status_array(
     result: RefractionDatumStaticsResult,
 ) -> np.ndarray:
@@ -5431,7 +4721,6 @@ def _receiver_manual_static_status_array(
         int(result.receiver_endpoint_key.shape[0]),
     )
 
-
 def _source_field_shift_s_array(result: RefractionDatumStaticsResult) -> np.ndarray:
     if not _has_field_correction_composition(result):
         return _disabled_field_float_array(int(result.source_endpoint_key.shape[0]))
@@ -5439,7 +4728,6 @@ def _source_field_shift_s_array(result: RefractionDatumStaticsResult) -> np.ndar
         result.source_field_shift_s,
         int(result.source_endpoint_key.shape[0]),
     )
-
 
 def _source_field_static_status_array(
     result: RefractionDatumStaticsResult,
@@ -5451,7 +4739,6 @@ def _source_field_static_status_array(
         int(result.source_endpoint_key.shape[0]),
     )
 
-
 def _receiver_field_shift_s_array(result: RefractionDatumStaticsResult) -> np.ndarray:
     if not _has_field_correction_composition(result):
         return _disabled_field_float_array(int(result.receiver_endpoint_key.shape[0]))
@@ -5459,7 +4746,6 @@ def _receiver_field_shift_s_array(result: RefractionDatumStaticsResult) -> np.nd
         result.receiver_field_shift_s,
         int(result.receiver_endpoint_key.shape[0]),
     )
-
 
 def _receiver_field_static_status_array(
     result: RefractionDatumStaticsResult,
@@ -5471,7 +4757,6 @@ def _receiver_field_static_status_array(
         int(result.receiver_endpoint_key.shape[0]),
     )
 
-
 def _source_field_shift_s_sorted_array(
     result: RefractionDatumStaticsResult,
 ) -> np.ndarray:
@@ -5481,7 +4766,6 @@ def _source_field_shift_s_sorted_array(
         result.source_field_shift_s_sorted,
         int(result.sorted_trace_index.shape[0]),
     )
-
 
 def _receiver_field_shift_s_sorted_array(
     result: RefractionDatumStaticsResult,
@@ -5493,7 +4777,6 @@ def _receiver_field_shift_s_sorted_array(
         int(result.sorted_trace_index.shape[0]),
     )
 
-
 def _trace_field_shift_s_sorted_array(
     result: RefractionDatumStaticsResult,
 ) -> np.ndarray:
@@ -5503,7 +4786,6 @@ def _trace_field_shift_s_sorted_array(
         result.trace_field_shift_s_sorted,
         int(result.sorted_trace_index.shape[0]),
     )
-
 
 def _trace_field_static_status_sorted_array(
     result: RefractionDatumStaticsResult,
@@ -5515,14 +4797,12 @@ def _trace_field_static_status_sorted_array(
         int(result.sorted_trace_index.shape[0]),
     )
 
-
 def _base_refraction_trace_shift_s_sorted_array(
     result: RefractionDatumStaticsResult,
 ) -> np.ndarray:
     if result.base_refraction_trace_shift_s_sorted is None:
         return _float_array(result.refraction_trace_shift_s_sorted)
     return _float_array(result.base_refraction_trace_shift_s_sorted)
-
 
 def _optional_field_float_array(value: object, shape: int) -> np.ndarray:
     if value is None:
@@ -5534,7 +4814,6 @@ def _optional_field_float_array(value: object, shape: int) -> np.ndarray:
         )
     return arr
 
-
 def _optional_field_status_array(value: object, shape: int) -> np.ndarray:
     if value is None:
         raise RefractionStaticArtifactError('field correction status array is missing')
@@ -5545,14 +4824,11 @@ def _optional_field_status_array(value: object, shape: int) -> np.ndarray:
         )
     return arr
 
-
 def _disabled_field_float_array(shape: int) -> np.ndarray:
     return np.zeros(int(shape), dtype=np.float64)
 
-
 def _disabled_field_status_array(shape: int) -> np.ndarray:
     return np.full(int(shape), _FIELD_DISABLED_STATUS, dtype='<U16')
-
 
 def _field_static_valid_mask(
     *,
@@ -5562,7 +4838,6 @@ def _field_static_valid_mask(
     status_text = np.asarray(status).astype(str)
     valid_status = np.isin(status_text, tuple(_FIELD_TOTAL_VALID_STATUSES))
     return np.ascontiguousarray(valid_status & np.isfinite(shift_s), dtype=bool)
-
 
 def _total_with_field_shift_s(
     *,
@@ -5586,7 +4861,6 @@ def _total_with_field_shift_s(
     out[valid] = refraction[valid] + field[valid]
     return np.ascontiguousarray(out, dtype=np.float64)
 
-
 def _final_trace_shift_s_sorted(
     result: RefractionDatumStaticsResult,
 ) -> np.ndarray:
@@ -5598,14 +4872,12 @@ def _final_trace_shift_s_sorted(
         field_status=_trace_field_static_status_sorted_array(result),
     )
 
-
 def _final_trace_static_status_sorted_array(
     result: RefractionDatumStaticsResult,
 ) -> np.ndarray:
     if result.final_trace_static_status_sorted is not None:
         return _string_array(result.final_trace_static_status_sorted)
     return _string_array(result.trace_static_status_sorted)
-
 
 def _final_trace_static_valid_mask_sorted_array(
     result: RefractionDatumStaticsResult,
@@ -5614,14 +4886,12 @@ def _final_trace_static_valid_mask_sorted_array(
         return _bool_array(result.final_trace_static_valid_mask_sorted)
     return _bool_array(result.trace_static_valid_mask_sorted)
 
-
 def _applied_field_shift_s_sorted_array(
     result: RefractionDatumStaticsResult,
 ) -> np.ndarray:
     if result.applied_field_shift_s_sorted is not None:
         return _float_array(result.applied_field_shift_s_sorted)
     return np.zeros(int(result.sorted_trace_index.shape[0]), dtype=np.float64)
-
 
 def _trace_endpoint_key_sorted_array(
     result: RefractionDatumStaticsResult,
@@ -5644,7 +4914,6 @@ def _trace_endpoint_key_sorted_array(
             f'{endpoint}_endpoint_key_sorted shape mismatch'
         )
     return out
-
 
 def _endpoint_shift_to_trace_order(
     *,
@@ -5678,7 +4947,6 @@ def _endpoint_shift_to_trace_order(
         out[index] = shift_by_key[text]
     return np.ascontiguousarray(out, dtype=np.float64)
 
-
 def _applied_endpoint_field_shift_s(
     *,
     field_shift_s: np.ndarray,
@@ -5694,7 +4962,6 @@ def _applied_endpoint_field_shift_s(
     out[valid] = field[valid]
     return np.ascontiguousarray(out, dtype=np.float64)
 
-
 def _component_qc_stats_ms(
     components_s: Mapping[str, np.ndarray],
 ) -> dict[str, dict[str, float | None]]:
@@ -5706,7 +4973,6 @@ def _component_qc_stats_ms(
         }
         for name, values in components_s.items()
     }
-
 
 from app.services.refraction_static_artifacts.field_corrections import (  # noqa: E402
     _applied_endpoint_field_shift_s,
@@ -5747,197 +5013,32 @@ from app.services.refraction_static_artifacts.field_corrections import (  # noqa
     _trace_field_static_status_sorted_array,
 )
 
-def _request_summary(req: RefractionStaticApplyRequest) -> dict[str, Any]:
-    return {
-        'file_id': req.file_id,
-        'key1_byte': int(req.key1_byte),
-        'key2_byte': int(req.key2_byte),
-        'pick_source_kind': req.pick_source.kind,
-        'model_method': req.model.method,
-        'apply_mode': req.apply.mode,
-        'register_corrected_file': bool(req.apply.register_corrected_file),
-    }
+def _request_summary(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_request_summary')(*args, **kwargs)
 
+def _layer_velocity_modes_for_request(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_layer_velocity_modes_for_request')(*args, **kwargs)
 
-def _layer_velocity_modes_for_request(
-    req: RefractionStaticApplyRequest,
-) -> dict[str, str]:
-    if req.model.method != 'multilayer_time_term':
-        return {}
-    return {
-        str(config.kind): str(config.velocity_mode)
-        for config in normalize_refraction_static_layers(req.model)
-    }
+def _request_has_cell_velocity_layer(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_request_has_cell_velocity_layer')(*args, **kwargs)
 
+def _request_cell_velocity_layer(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_request_cell_velocity_layer')(*args, **kwargs)
 
-def _request_has_cell_velocity_layer(req: RefractionStaticApplyRequest) -> bool:
-    return bool(_request_cell_velocity_layer_kinds(req))
+def _result_has_cell_velocity_arrays(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_result_has_cell_velocity_arrays')(*args, **kwargs)
 
+def _cell_velocity_artifact_result_for_layer(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_cell_velocity_artifact_result_for_layer')(*args, **kwargs)
 
-def _request_cell_velocity_layer(
-    req: RefractionStaticApplyRequest,
-    *,
-    layer_kind: RefractionLayerKind | None = None,
-) -> Any | None:
-    if req.model.method != 'multilayer_time_term':
-        return None
-    for layer in normalize_refraction_static_layers(req.model):
-        if layer.velocity_mode != 'solve_cell':
-            continue
-        if layer_kind is None or layer.kind == layer_kind:
-            return layer
-    return None
+def _cell_velocity_layer_result(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_cell_velocity_layer_result')(*args, **kwargs)
 
+def _cell_velocity_artifact_result_from_layer(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_cell_velocity_artifact_result_from_layer')(*args, **kwargs)
 
-def _result_has_cell_velocity_arrays(result: RefractionDatumStaticsResult) -> bool:
-    return (
-        result.active_cell_id is not None
-        and result.inactive_cell_id is not None
-        and result.cell_bedrock_velocity_m_s is not None
-        and result.cell_bedrock_slowness_s_per_m is not None
-        and result.cell_velocity_status is not None
-    )
-
-
-def _cell_velocity_artifact_result_for_layer(
-    *,
-    result: RefractionDatumStaticsResult,
-    req: RefractionStaticApplyRequest,
-    layer_kind: RefractionLayerKind,
-) -> RefractionDatumStaticsResult:
-    if req.model.method != 'multilayer_time_term':
-        return result
-
-    layer = _cell_velocity_layer_result(result, layer_kind)
-    if layer is None:
-        if (
-            len(_request_cell_velocity_layer_kinds(req)) == 1
-            and _result_has_cell_velocity_arrays(result)
-        ):
-            return result
-        raise RefractionStaticArtifactError(
-            f'{layer_kind} cell velocity artifacts require layer solve result arrays'
-        )
-
-    return _cell_velocity_artifact_result_from_layer(
-        result=result,
-        req=req,
-        layer=layer,
-    )
-
-
-def _cell_velocity_layer_result(
-    result: RefractionDatumStaticsResult,
-    layer_kind: RefractionLayerKind,
-) -> RefractionLayerSolveResult | None:
-    for layer in result.layer_results or ():
-        if layer.layer_kind == layer_kind:
-            return layer
-    return None
-
-
-def _cell_velocity_artifact_result_from_layer(
-    *,
-    result: RefractionDatumStaticsResult,
-    req: RefractionStaticApplyRequest,
-    layer: RefractionLayerSolveResult,
-) -> RefractionDatumStaticsResult:
-    n_rows = int(result.row_trace_index_sorted.shape[0])
-    active_cell_id = _required_layer_cell_id_array(
-        layer.active_cell_id,
-        name=f'{layer.layer_kind}.active_cell_id',
-    )
-    inactive_cell_id = _required_layer_cell_id_array(
-        layer.inactive_cell_id,
-        name=f'{layer.layer_kind}.inactive_cell_id',
-    )
-    n_total_cells = _cell_velocity_n_total_cells(req)
-    cell_velocity = _layer_cell_values_for_active_cells(
-        layer.cell_velocity_m_s,
-        active_cell_id=active_cell_id,
-        n_total_cells=n_total_cells,
-        name=f'{layer.layer_kind}.cell_velocity_m_s',
-    )
-    cell_slowness = _layer_cell_values_for_active_cells(
-        layer.cell_slowness_s_per_m,
-        active_cell_id=active_cell_id,
-        n_total_cells=n_total_cells,
-        name=f'{layer.layer_kind}.cell_slowness_s_per_m',
-    )
-    cell_status = _layer_cell_status_for_active_cells(
-        layer.cell_velocity_status,
-        active_cell_id=active_cell_id,
-        n_total_cells=n_total_cells,
-        name=f'{layer.layer_kind}.cell_velocity_status',
-    )
-    velocity_summary = _stat(cell_velocity, 'median')
-    slowness_summary = _stat(cell_slowness, 'median')
-    qc = {**result.qc, **layer.qc}
-    if 'layers' in result.qc:
-        qc['layers'] = result.qc['layers']
-    return replace(
-        result,
-        bedrock_velocity_mode='solve_cell',
-        bedrock_velocity_m_s=(
-            result.bedrock_velocity_m_s
-            if velocity_summary is None
-            else velocity_summary
-        ),
-        bedrock_slowness_s_per_m=(
-            result.bedrock_slowness_s_per_m
-            if slowness_summary is None
-            else slowness_summary
-        ),
-        active_cell_id=active_cell_id,
-        inactive_cell_id=inactive_cell_id,
-        cell_bedrock_velocity_m_s=cell_velocity,
-        cell_bedrock_slowness_s_per_m=cell_slowness,
-        cell_velocity_status=cell_status,
-        row_midpoint_cell_id=_row_midpoint_cell_id_for_cell_velocity_layer(
-            result=result,
-            req=req,
-            layer=layer,
-            n_rows=n_rows,
-        ),
-        modeled_pick_time_s=_layer_trace_float_array(
-            layer.trace_predicted_time_s_sorted,
-            n_rows=n_rows,
-            name=f'{layer.layer_kind}.trace_predicted_time_s_sorted',
-        ),
-        residual_time_s=_layer_trace_float_array(
-            layer.trace_residual_s_sorted,
-            n_rows=n_rows,
-            name=f'{layer.layer_kind}.trace_residual_s_sorted',
-        ),
-        used_row_mask=_layer_trace_bool_array(
-            layer.used_observation_mask_sorted,
-            n_rows=n_rows,
-            name=f'{layer.layer_kind}.used_observation_mask_sorted',
-        ),
-        rejected_by_robust_mask=_layer_trace_bool_array(
-            (
-                np.zeros(n_rows, dtype=bool)
-                if layer.rejected_by_robust_mask_sorted is None
-                else layer.rejected_by_robust_mask_sorted
-            ),
-            n_rows=n_rows,
-            name=f'{layer.layer_kind}.rejected_by_robust_mask_sorted',
-        ),
-        qc=qc,
-    )
-
-
-def _cell_velocity_n_total_cells(req: RefractionStaticApplyRequest) -> int:
-    refractor_cell = req.model.refractor_cell
-    if refractor_cell is None:
-        raise RefractionStaticArtifactError(
-            'model.refractor_cell is required for cell velocity artifacts'
-        )
-    grid = build_refraction_cell_grid(
-        effective_refraction_cell_grid_config(refractor_cell)
-    )
-    return int(grid.cell_id.shape[0])
-
+def _cell_velocity_n_total_cells(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_cell_velocity_n_total_cells')(*args, **kwargs)
 
 def _required_layer_cell_id_array(value: object, *, name: str) -> np.ndarray:
     if value is None:
@@ -5947,592 +5048,84 @@ def _required_layer_cell_id_array(value: object, *, name: str) -> np.ndarray:
         raise RefractionStaticArtifactError(f'{name} must be one-dimensional')
     return np.ascontiguousarray(array, dtype=np.int64)
 
+def _layer_cell_values_for_active_cells(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_layer_cell_values_for_active_cells')(*args, **kwargs)
 
-def _layer_cell_values_for_active_cells(
-    value: object,
-    *,
-    active_cell_id: np.ndarray,
-    n_total_cells: int,
-    name: str,
-) -> np.ndarray:
-    if value is None:
-        raise RefractionStaticArtifactError(f'{name} is required')
-    array = np.asarray(value, dtype=np.float64)
-    if array.ndim != 1:
-        raise RefractionStaticArtifactError(f'{name} must be one-dimensional')
-    if array.shape == active_cell_id.shape:
-        return np.ascontiguousarray(array, dtype=np.float64)
-    if array.shape == (int(n_total_cells),):
-        return np.ascontiguousarray(array[active_cell_id], dtype=np.float64)
-    raise RefractionStaticArtifactError(
-        f'{name} must match active_cell_id length or total cell count'
-    )
+def _layer_cell_status_for_active_cells(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_layer_cell_status_for_active_cells')(*args, **kwargs)
 
+def _layer_trace_float_array(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_layer_trace_float_array')(*args, **kwargs)
 
-def _layer_cell_status_for_active_cells(
-    value: object,
-    *,
-    active_cell_id: np.ndarray,
-    n_total_cells: int,
-    name: str,
-) -> np.ndarray:
-    if value is None:
-        return _string_array(np.full(active_cell_id.shape, 'solved', dtype='<U6'))
-    status = _string_array(value)
-    if status.shape == active_cell_id.shape:
-        _validate_status_array(status, name=name)
-        return status
-    if status.shape == (int(n_total_cells),):
-        out = _string_array(status[active_cell_id])
-        _validate_status_array(out, name=name)
-        return out
-    raise RefractionStaticArtifactError(
-        f'{name} must match active_cell_id length or total cell count'
-    )
+def _layer_trace_bool_array(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_layer_trace_bool_array')(*args, **kwargs)
 
+def _row_midpoint_cell_id_for_cell_velocity_layer(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_row_midpoint_cell_id_for_cell_velocity_layer')(*args, **kwargs)
 
-def _layer_trace_float_array(value: object, *, n_rows: int, name: str) -> np.ndarray:
-    array = np.asarray(value, dtype=np.float64)
-    if array.shape != (int(n_rows),):
-        raise RefractionStaticArtifactError(
-            f'{name} length must match residual rows'
-        )
-    return np.ascontiguousarray(array, dtype=np.float64)
+def _compute_row_midpoint_cell_id(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_compute_row_midpoint_cell_id')(*args, **kwargs)
 
+def _row_endpoint_xy(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_row_endpoint_xy')(*args, **kwargs)
 
-def _layer_trace_bool_array(value: object, *, n_rows: int, name: str) -> np.ndarray:
-    array = np.asarray(value, dtype=bool)
-    if array.shape != (int(n_rows),):
-        raise RefractionStaticArtifactError(
-            f'{name} length must match residual rows'
-        )
-    return np.ascontiguousarray(array, dtype=bool)
+def _cell_velocity_candidate_row_mask(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_cell_velocity_candidate_row_mask')(*args, **kwargs)
 
+def _cell_velocity_component(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_cell_velocity_component')(*args, **kwargs)
 
-def _row_midpoint_cell_id_for_cell_velocity_layer(
-    *,
-    result: RefractionDatumStaticsResult,
-    req: RefractionStaticApplyRequest,
-    layer: RefractionLayerSolveResult,
-    n_rows: int,
-) -> np.ndarray:
-    if layer.row_midpoint_cell_id is not None:
-        array = np.asarray(layer.row_midpoint_cell_id, dtype=np.int64)
-        if array.shape == (int(n_rows),):
-            return np.ascontiguousarray(array, dtype=np.int64)
-    if result.row_midpoint_cell_id is not None:
-        array = np.asarray(result.row_midpoint_cell_id, dtype=np.int64)
-        if array.shape == (int(n_rows),):
-            return np.ascontiguousarray(array, dtype=np.int64)
-    return _compute_row_midpoint_cell_id(result=result, req=req, n_rows=n_rows)
+def _cell_velocity_layer_bounds(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_cell_velocity_layer_bounds')(*args, **kwargs)
 
+def _cell_velocity_min_observations_per_cell(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_cell_velocity_min_observations_per_cell')(*args, **kwargs)
 
-def _compute_row_midpoint_cell_id(
-    *,
-    result: RefractionDatumStaticsResult,
-    req: RefractionStaticApplyRequest,
-    n_rows: int,
-) -> np.ndarray:
-    refractor_cell = req.model.refractor_cell
-    if refractor_cell is None:
-        raise RefractionStaticArtifactError(
-            'model.refractor_cell is required for cell velocity artifacts'
-        )
-    source_x, source_y = _row_endpoint_xy(result, endpoint='source', n_rows=n_rows)
-    receiver_x, receiver_y = _row_endpoint_xy(
-        result,
-        endpoint='receiver',
-        n_rows=n_rows,
-    )
-    source_projected = project_refraction_cell_points(
-        x_m=source_x,
-        y_m=source_y,
-        mode=refractor_cell.coordinate_mode,
-        line_origin_x_m=refractor_cell.line_origin_x_m,
-        line_origin_y_m=refractor_cell.line_origin_y_m,
-        line_azimuth_deg=refractor_cell.line_azimuth_deg,
-    )
-    receiver_projected = project_refraction_cell_points(
-        x_m=receiver_x,
-        y_m=receiver_y,
-        mode=refractor_cell.coordinate_mode,
-        line_origin_x_m=refractor_cell.line_origin_x_m,
-        line_origin_y_m=refractor_cell.line_origin_y_m,
-        line_azimuth_deg=refractor_cell.line_azimuth_deg,
-    )
-    grid = build_refraction_cell_grid(
-        effective_refraction_cell_grid_config(refractor_cell)
-    )
-    assignment = assign_observation_midpoint_cells(
-        grid,
-        source_x_m=source_projected.x_m,
-        source_y_m=source_projected.y_m,
-        receiver_x_m=receiver_projected.x_m,
-        receiver_y_m=receiver_projected.y_m,
-    )
-    return np.ascontiguousarray(assignment.cell_id, dtype=np.int64)
+def _refractor_velocity_cell_rows(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_refractor_velocity_cell_rows')(*args, **kwargs)
 
+def _cell_solver_history_csv_row(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_cell_solver_history_csv_row')(*args, **kwargs)
 
-def _row_endpoint_xy(
-    result: RefractionDatumStaticsResult,
-    *,
-    endpoint: str,
-    n_rows: int,
-) -> tuple[np.ndarray, np.ndarray]:
-    if endpoint == 'source':
-        row_key = result.row_source_endpoint_key
-        endpoint_key = result.source_endpoint_key
-        x_m = result.source_x_m
-        y_m = result.source_y_m
-    elif endpoint == 'receiver':
-        row_key = result.row_receiver_endpoint_key
-        endpoint_key = result.receiver_endpoint_key
-        x_m = result.receiver_x_m
-        y_m = result.receiver_y_m
-    else:
-        raise RefractionStaticArtifactError(f'unsupported endpoint: {endpoint}')
-    if row_key is None:
-        raise RefractionStaticArtifactError(
-            f'row_{endpoint}_endpoint_key is required for layer cell artifacts'
-        )
-    row_key_array = np.asarray(row_key, dtype=object)
-    if row_key_array.shape != (int(n_rows),):
-        raise RefractionStaticArtifactError(
-            f'row_{endpoint}_endpoint_key length must match residual rows'
-        )
-    endpoint_key_array = np.asarray(endpoint_key, dtype=object)
-    x_array = np.asarray(x_m, dtype=np.float64)
-    y_array = np.asarray(y_m, dtype=np.float64)
-    if endpoint_key_array.shape != x_array.shape or x_array.shape != y_array.shape:
-        raise RefractionStaticArtifactError(
-            f'{endpoint} endpoint key and coordinate arrays must match'
-        )
-    index_by_key = {
-        str(key): index for index, key in enumerate(endpoint_key_array.tolist())
-    }
-    indices: list[int] = []
-    for key in row_key_array.tolist():
-        try:
-            indices.append(index_by_key[str(key)])
-        except KeyError as exc:
-            raise RefractionStaticArtifactError(
-                f'row {endpoint} endpoint key {key!r} is missing from endpoint table'
-            ) from exc
-    index_array = np.asarray(indices, dtype=np.int64)
-    return (
-        np.ascontiguousarray(x_array[index_array], dtype=np.float64),
-        np.ascontiguousarray(y_array[index_array], dtype=np.float64),
-    )
+def _cell_solver_history_cell_counts(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_cell_solver_history_cell_counts')(*args, **kwargs)
 
+def _history_residual_stats_ms(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_history_residual_stats_ms')(*args, **kwargs)
 
-def _cell_velocity_candidate_row_mask(
-    result: RefractionDatumStaticsResult,
-    req: RefractionStaticApplyRequest,
-    *,
-    layer_kind: RefractionLayerKind | None = None,
-) -> np.ndarray:
-    n_rows = int(result.row_trace_index_sorted.shape[0])
-    layer = _request_cell_velocity_layer(req, layer_kind=layer_kind)
-    if layer is None:
-        return np.ones(n_rows, dtype=bool)
-    distance = np.asarray(result.row_distance_m, dtype=np.float64)
-    if distance.shape != (n_rows,):
-        raise RefractionStaticArtifactError(
-            'row_distance_m length must match residual rows'
-        )
-    valid_observation = np.asarray(result.valid_observation_mask_sorted, dtype=bool)
-    if valid_observation.shape != (n_rows,):
-        raise RefractionStaticArtifactError(
-            'valid_observation_mask_sorted length must match residual rows'
-        )
-    try:
-        masks = build_refraction_layer_observation_masks_from_arrays(
-            base_valid_mask_sorted=valid_observation,
-            offset_m_sorted=distance,
-            rejection_reason_sorted=np.full(n_rows, 'ok', dtype='<U32'),
-            model=req.model,
-        )
-        mask = masks.layer_used_mask_sorted[layer.kind]
-    except (KeyError, ValueError) as exc:
-        raise RefractionStaticArtifactError(
-            f'could not build cell velocity layer observation mask for {layer.kind}'
-        ) from exc
-    return np.ascontiguousarray(mask, dtype=bool)
+def _history_max_abs(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_history_max_abs')(*args, **kwargs)
 
+def _initial_cell_v2_m_s(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_initial_cell_v2_m_s')(*args, **kwargs)
 
-def _cell_velocity_component(layer_kind: RefractionLayerKind) -> str:
-    try:
-        return _CELL_VELOCITY_COMPONENT_BY_LAYER[layer_kind]
-    except KeyError as exc:
-        raise RefractionStaticArtifactError(
-            f'unsupported cell velocity layer kind: {layer_kind}'
-        ) from exc
+def _history_smoothing_weight(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_history_smoothing_weight')(*args, **kwargs)
 
+def _history_robust_iteration_count(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_history_robust_iteration_count')(*args, **kwargs)
 
-def _cell_velocity_layer_bounds(
-    req: RefractionStaticApplyRequest,
-    layer: Any,
-) -> tuple[float | None, float | None]:
-    min_velocity = getattr(layer, 'min_velocity_m_s', None)
-    max_velocity = getattr(layer, 'max_velocity_m_s', None)
-    if layer.kind == 'v2_t1':
-        if min_velocity is None:
-            min_velocity = req.model.min_bedrock_velocity_m_s
-        if max_velocity is None:
-            max_velocity = req.model.max_bedrock_velocity_m_s
-    return min_velocity, max_velocity
+def _history_convergence_reason(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_history_convergence_reason')(*args, **kwargs)
 
+def _refractor_cell_center_alias_arrays(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_refractor_cell_center_alias_arrays')(*args, **kwargs)
 
-def _cell_velocity_min_observations_per_cell(
-    req: RefractionStaticApplyRequest,
-    *,
-    layer_kind: RefractionLayerKind | None = None,
-) -> int:
-    layer = _request_cell_velocity_layer(req, layer_kind=layer_kind)
-    if layer is not None and layer.min_observations_per_cell is not None:
-        return int(layer.min_observations_per_cell)
-    refractor_cell = req.model.refractor_cell
-    if refractor_cell is None:
-        raise RefractionStaticArtifactError(
-            'model.refractor_cell is required for cell velocity artifacts'
-        )
-    return int(refractor_cell.min_observations_per_cell)
+def _unique_observation_endpoint_count_by_cell(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_unique_observation_endpoint_count_by_cell')(*args, **kwargs)
 
-
-def _refractor_velocity_cell_rows(
-    arrays: dict[str, np.ndarray],
-) -> list[dict[str, object]]:
-    rows: list[dict[str, object]] = []
-    n_cells = int(arrays['cell_id'].shape[0])
-    for index in range(n_cells):
-        rows.append(
-            {
-                'cell_id': int(arrays['cell_id'][index]),
-                'ix': int(arrays['ix'][index]),
-                'iy': int(arrays['iy'][index]),
-                'cell_ix': int(arrays['cell_ix'][index]),
-                'cell_iy': int(arrays['cell_iy'][index]),
-                'coordinate_mode': str(arrays['coordinate_mode'][index]),
-                'x_min_m': _csv_grid_float(arrays['x_min_m'][index]),
-                'x_max_m': _csv_grid_float(arrays['x_max_m'][index]),
-                'y_min_m': _csv_grid_float(arrays['y_min_m'][index]),
-                'y_max_m': _csv_grid_float(arrays['y_max_m'][index]),
-                'x_center_m': _csv_grid_float(arrays['x_center_m'][index]),
-                'y_center_m': _csv_grid_float(arrays['y_center_m'][index]),
-                'cell_center_x_m': _csv_grid_float(
-                    arrays['cell_center_x_m'][index]
-                ),
-                'cell_center_y_m': _csv_grid_float(
-                    arrays['cell_center_y_m'][index]
-                ),
-                'cell_center_inline_m': _csv_grid_float(
-                    arrays['cell_center_inline_m'][index]
-                ),
-                'cell_center_crossline_m': _csv_grid_float(
-                    arrays['cell_center_crossline_m'][index]
-                ),
-                'active': _csv_bool(arrays['active_cell_mask'][index]),
-                'n_observations': int(
-                    arrays['n_observations_per_cell'][index]
-                ),
-                'n_used_observations': int(
-                    arrays['n_used_observations_per_cell'][index]
-                ),
-                'n_rejected_observations': int(
-                    arrays['n_rejected_observations_per_cell'][index]
-                ),
-                'n_sources': int(arrays['n_sources_per_cell'][index]),
-                'n_receivers': int(arrays['n_receivers_per_cell'][index]),
-                'cell_velocity_layer_kind': str(
-                    arrays['cell_velocity_layer_kind'][index]
-                ),
-                'cell_velocity_component': str(
-                    arrays['cell_velocity_component'][index]
-                ),
-                'velocity_m_s': _csv_float(arrays['velocity_m_s'][index]),
-                'v2_m_s': _csv_float(arrays['v2_m_s'][index]),
-                'slowness_s_per_m': _csv_float(
-                    arrays['slowness_s_per_m'][index]
-                ),
-                'initial_velocity_m_s': _csv_float(
-                    arrays['initial_velocity_m_s'][index]
-                ),
-                'initial_v2_m_s': _csv_float(arrays['initial_v2_m_s'][index]),
-                'velocity_update_from_initial_m_s': _csv_float(
-                    arrays['velocity_update_from_initial_m_s'][index]
-                ),
-                'v2_update_from_initial_m_s': _csv_float(
-                    arrays['v2_update_from_initial_m_s'][index]
-                ),
-                'velocity_status': str(arrays['velocity_status'][index]),
-                'status_reason': str(arrays['status_reason'][index]),
-                'residual_rms_ms': _csv_float(arrays['residual_rms_ms'][index]),
-                'residual_mad_ms': _csv_float(arrays['residual_mad_ms'][index]),
-                'residual_mean_ms': _csv_float(arrays['residual_mean_ms'][index]),
-                'residual_p95_abs_ms': _csv_float(
-                    arrays['residual_p95_abs_ms'][index]
-                ),
-                'smoothing_enabled': _csv_bool(arrays['smoothing_enabled'][index]),
-                'smoothing_weight': _csv_float(arrays['smoothing_weight'][index]),
-                'smoothing_neighbor_count': int(
-                    arrays['smoothing_neighbor_count'][index]
-                ),
-            }
-        )
-    return rows
-
-
-def _cell_solver_history_csv_row(
-    row: RefractionCellSolverHistoryRow,
-) -> dict[str, object]:
-    return {
-        'iteration': int(row.iteration),
-        'stage': row.stage,
-        'n_candidate_observations': int(row.n_candidate_observations),
-        'n_used_observations': int(row.n_used_observations),
-        'n_rejected_observations': int(row.n_rejected_observations),
-        'n_active_cells': int(row.n_active_cells),
-        'n_low_fold_cells': int(row.n_low_fold_cells),
-        'n_empty_cells': int(row.n_empty_cells),
-        'residual_rms_ms': _csv_float(row.residual_rms_ms),
-        'residual_mad_ms': _csv_float(row.residual_mad_ms),
-        'max_abs_residual_ms': _csv_float(row.max_abs_residual_ms),
-        'median_velocity_m_s': _csv_float(row.median_velocity_m_s),
-        'median_v2_m_s': _csv_float(row.median_v2_m_s),
-        'min_velocity_m_s': _csv_float(row.min_velocity_m_s),
-        'min_v2_m_s': _csv_float(row.min_v2_m_s),
-        'max_velocity_m_s': _csv_float(row.max_velocity_m_s),
-        'max_v2_m_s': _csv_float(row.max_v2_m_s),
-        'max_abs_velocity_update_m_s': _csv_float(
-            row.max_abs_velocity_update_m_s
-        ),
-        'max_abs_v2_update_m_s': _csv_float(row.max_abs_v2_update_m_s),
-        'smoothing_weight': _csv_float(row.smoothing_weight),
-        'damping_weight': _csv_float(row.damping_weight),
-        'robust_threshold': _csv_float(row.robust_threshold),
-        'converged': _csv_bool(row.converged),
-        'convergence_reason': row.convergence_reason,
-    }
-
-
-def _cell_solver_history_cell_counts(
-    arrays: dict[str, np.ndarray],
-) -> dict[str, int]:
-    active_mask = np.asarray(arrays['active_cell_mask'], dtype=bool)
-    velocity_status = np.asarray(arrays['velocity_status']).astype(str, copy=False)
-    n_observations = np.asarray(arrays['n_observations_per_cell'], dtype=np.int64)
-    return {
-        'active': int(np.count_nonzero(active_mask)),
-        'low_fold': int(
-            np.count_nonzero(velocity_status == LOW_FOLD_CELL_VELOCITY_STATUS)
-        ),
-        'empty': int(np.count_nonzero(n_observations == 0)),
-    }
-
-
-def _history_residual_stats_ms(
-    result: RefractionDatumStaticsResult,
-    *,
-    used_row_mask: np.ndarray,
-) -> dict[str, float | None]:
-    residual_ms = (
-        np.asarray(result.residual_time_s, dtype=np.float64)[
-            np.asarray(used_row_mask, dtype=bool)
-        ]
-        * 1000.0
-    )
-    return {
-        'rms': _residual_stat(residual_ms, 'rms'),
-        'mad': _residual_stat(residual_ms, 'mad'),
-        'max_abs': _residual_stat(residual_ms, 'max_abs'),
-    }
-
-
-def _history_max_abs(values: np.ndarray) -> float | None:
-    finite = np.asarray(values, dtype=np.float64)
-    finite = finite[np.isfinite(finite)]
-    if finite.size == 0:
-        return None
-    return float(np.max(np.abs(finite)))
-
-
-def _initial_cell_v2_m_s(
-    req: RefractionStaticApplyRequest,
-    *,
-    layer_kind: RefractionLayerKind | None = None,
-) -> float:
-    layer = _request_cell_velocity_layer(req, layer_kind=layer_kind)
-    if layer is not None:
-        value = getattr(layer, 'initial_velocity_m_s', None)
-        if value is not None:
-            return float(value)
-        if layer.kind == 'v2_t1' and req.model.initial_bedrock_velocity_m_s is not None:
-            return float(req.model.initial_bedrock_velocity_m_s)
-        min_velocity, max_velocity = _cell_velocity_layer_bounds(req, layer)
-        if min_velocity is not None and max_velocity is not None:
-            return 0.5 * (float(min_velocity) + float(max_velocity))
-        raise RefractionStaticArtifactError(
-            'cell velocity layer initial_velocity_m_s is required'
-        )
-
-    value = req.model.initial_bedrock_velocity_m_s
-    if value is not None:
-        return float(value)
-    return 0.5 * (
-        float(req.model.min_bedrock_velocity_m_s)
-        + float(req.model.max_bedrock_velocity_m_s)
-    )
-
-
-def _history_smoothing_weight(
-    req: RefractionStaticApplyRequest,
-    *,
-    layer_kind: RefractionLayerKind | None = None,
-) -> float:
-    layer = _request_cell_velocity_layer(req, layer_kind=layer_kind)
-    if layer is not None and layer.smoothing_weight is not None:
-        return float(layer.smoothing_weight)
-    refractor_cell = req.model.refractor_cell
-    if refractor_cell is None:
-        raise RefractionStaticArtifactError(
-            'model.refractor_cell is required for cell solver history'
-        )
-    return float(refractor_cell.velocity_smoothing_weight)
-
-
-def _history_robust_iteration_count(
-    result: RefractionDatumStaticsResult,
-    req: RefractionStaticApplyRequest,
-) -> int:
-    raw = result.qc.get('robust_iteration_count')
-    if raw is not None:
-        return int(raw)
-    if req.solver.robust.enabled and np.count_nonzero(result.rejected_by_robust_mask):
-        return 1
-    return 0
-
-
-def _history_convergence_reason(
-    *,
-    robust_iteration_count: int,
-    n_robust_rejected_observations: int,
-    smoothing_weight: float,
-) -> str:
-    if robust_iteration_count > 0 or n_robust_rejected_observations > 0:
-        return 'robust_reweight_converged'
-    if smoothing_weight > 0.0:
-        return 'smoothed_least_squares_converged'
-    return 'least_squares_converged'
-
-
-def _refractor_cell_center_alias_arrays(
-    *,
-    x_center_m: np.ndarray,
-    y_center_m: np.ndarray,
-    coordinate_mode: str,
-    line_origin_x_m: float | None,
-    line_origin_y_m: float | None,
-    line_azimuth_deg: float | None,
-) -> dict[str, np.ndarray]:
-    x_center = np.asarray(x_center_m, dtype=np.float64)
-    y_center = np.asarray(y_center_m, dtype=np.float64)
-    if x_center.shape != y_center.shape:
-        raise RefractionStaticArtifactError(
-            'refractor cell center coordinate shape mismatch'
-        )
-
-    inline = np.full(x_center.shape, np.nan, dtype=np.float64)
-    crossline = np.full(x_center.shape, np.nan, dtype=np.float64)
-    center_x = np.ascontiguousarray(x_center, dtype=np.float64)
-    center_y = np.ascontiguousarray(y_center, dtype=np.float64)
-    if coordinate_mode == 'line_2d_projected':
-        origin_x = _required_finite_float(
-            line_origin_x_m,
-            name='model.refractor_cell.line_origin_x_m',
-        )
-        origin_y = _required_finite_float(
-            line_origin_y_m,
-            name='model.refractor_cell.line_origin_y_m',
-        )
-        azimuth_deg = _required_finite_float(
-            line_azimuth_deg,
-            name='model.refractor_cell.line_azimuth_deg',
-        )
-        inline = np.ascontiguousarray(x_center, dtype=np.float64)
-        crossline = np.ascontiguousarray(y_center, dtype=np.float64)
-        azimuth_rad = np.deg2rad(azimuth_deg)
-        inline_unit_x = float(np.sin(azimuth_rad))
-        inline_unit_y = float(np.cos(azimuth_rad))
-        center_x = np.ascontiguousarray(
-            origin_x + inline * inline_unit_x + crossline * inline_unit_y,
-            dtype=np.float64,
-        )
-        center_y = np.ascontiguousarray(
-            origin_y + inline * inline_unit_y - crossline * inline_unit_x,
-            dtype=np.float64,
-        )
-
-    return {
-        'x_m': center_x,
-        'y_m': center_y,
-        'inline_m': np.ascontiguousarray(inline, dtype=np.float64),
-        'crossline_m': np.ascontiguousarray(crossline, dtype=np.float64),
-    }
-
-
-def _unique_observation_endpoint_count_by_cell(
-    *,
-    row_midpoint_cell_id: np.ndarray,
-    endpoint_id: np.ndarray,
-    n_total_cells: int,
-) -> np.ndarray:
-    row_cell = np.asarray(row_midpoint_cell_id, dtype=np.int64)
-    endpoint = np.asarray(endpoint_id, dtype=np.int64)
-    if row_cell.shape != endpoint.shape:
-        raise RefractionStaticArtifactError(
-            'row endpoint arrays must match row_midpoint_cell_id shape'
-        )
-    counts = np.zeros(int(n_total_cells), dtype=np.int64)
-    for cell_id in range(int(n_total_cells)):
-        in_cell = row_cell == cell_id
-        if np.any(in_cell):
-            counts[cell_id] = int(np.unique(endpoint[in_cell]).shape[0])
-    return np.ascontiguousarray(counts, dtype=np.int64)
-
-
-def _refractor_cell_status_reasons(
-    *,
-    velocity_status: np.ndarray,
-    n_observations: np.ndarray,
-) -> np.ndarray:
-    status = np.asarray(velocity_status).astype(str, copy=False)
-    counts = np.asarray(n_observations, dtype=np.int64)
-    if status.shape != counts.shape:
-        raise RefractionStaticArtifactError(
-            'velocity_status and n_observations shape mismatch'
-        )
-    reasons: list[str] = []
-    for value, count in zip(status.tolist(), counts.tolist(), strict=True):
-        if value == LOW_FOLD_CELL_VELOCITY_STATUS:
-            reasons.append(LOW_FOLD_CELL_REJECTION_REASON)
-        elif value == 'inactive':
-            reasons.append('no_observations' if int(count) == 0 else 'inactive_cell')
-        else:
-            reasons.append(str(value))
-    return _string_array(reasons)
-
+def _refractor_cell_status_reasons(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_refractor_cell_status_reasons')(*args, **kwargs)
 
 def _required_cell_int_array(value: object, *, name: str) -> np.ndarray:
     if value is None:
         raise RefractionStaticArtifactError(f'solve_cell result requires {name}')
     return _int_array(value)
 
-
 def _required_cell_float_array(value: object, *, name: str) -> np.ndarray:
     if value is None:
         raise RefractionStaticArtifactError(f'solve_cell result requires {name}')
     return _float_array(value)
-
 
 def _required_cell_status_array(value: object, *, name: str) -> np.ndarray:
     if value is None:
@@ -6540,7 +5133,6 @@ def _required_cell_status_array(value: object, *, name: str) -> np.ndarray:
     status = _string_array(value)
     _validate_status_array(status, name=name)
     return status
-
 
 def _validate_refractor_velocity_cell_ids(
     *,
@@ -6565,91 +5157,14 @@ def _validate_refractor_velocity_cell_ids(
             f'missing={missing}, extra={extra}'
         )
 
+def _active_neighbor_count_by_cell(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_active_neighbor_count_by_cell')(*args, **kwargs)
 
-def _active_neighbor_count_by_cell(
-    *,
-    active_cell_id: np.ndarray,
-    n_total_cells: int,
-    number_of_cell_x: int,
-    number_of_cell_y: int,
-) -> np.ndarray:
-    active = {int(value) for value in np.asarray(active_cell_id).tolist()}
-    counts = np.zeros(int(n_total_cells), dtype=np.int64)
-    for cell_id in active:
-        ix = cell_id % int(number_of_cell_x)
-        iy = cell_id // int(number_of_cell_x)
-        neighbors = []
-        if ix > 0:
-            neighbors.append(cell_id - 1)
-        if ix + 1 < int(number_of_cell_x):
-            neighbors.append(cell_id + 1)
-        if iy > 0:
-            neighbors.append(cell_id - int(number_of_cell_x))
-        if iy + 1 < int(number_of_cell_y):
-            neighbors.append(cell_id + int(number_of_cell_x))
-        counts[cell_id] = sum(1 for neighbor in neighbors if neighbor in active)
-    return np.ascontiguousarray(counts, dtype=np.int64)
+def _estimated_cell_smoothing_rows(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_estimated_cell_smoothing_rows')(*args, **kwargs)
 
-
-def _estimated_cell_smoothing_rows(
-    *,
-    active_cell_mask: np.ndarray,
-    number_of_cell_x: int,
-    number_of_cell_y: int,
-    velocity_smoothing_weight: float,
-) -> int:
-    if float(velocity_smoothing_weight) == 0.0:
-        return 0
-    active = {
-        index
-        for index, enabled in enumerate(np.asarray(active_cell_mask, dtype=bool))
-        if bool(enabled)
-    }
-    n_edges = 0
-    for cell_id in active:
-        ix = cell_id % int(number_of_cell_x)
-        iy = cell_id // int(number_of_cell_x)
-        right = cell_id + 1
-        down = cell_id + int(number_of_cell_x)
-        if ix + 1 < int(number_of_cell_x) and right in active:
-            n_edges += 1
-        if iy + 1 < int(number_of_cell_y) and down in active:
-            n_edges += 1
-    return n_edges
-
-
-def _per_cell_residual_stats_ms(
-    *,
-    row_midpoint_cell_id: np.ndarray,
-    residual_time_s: np.ndarray,
-    used_row_mask: np.ndarray,
-    n_total_cells: int,
-) -> dict[str, np.ndarray]:
-    row_cell = np.asarray(row_midpoint_cell_id, dtype=np.int64)
-    residual_ms = np.asarray(residual_time_s, dtype=np.float64) * 1000.0
-    used = np.asarray(used_row_mask, dtype=bool)
-    out = {
-        'rms': np.full(int(n_total_cells), np.nan, dtype=np.float64),
-        'mad': np.full(int(n_total_cells), np.nan, dtype=np.float64),
-        'mean': np.full(int(n_total_cells), np.nan, dtype=np.float64),
-        'p95_abs': np.full(int(n_total_cells), np.nan, dtype=np.float64),
-    }
-    for cell_id in range(int(n_total_cells)):
-        values = residual_ms[(row_cell == cell_id) & used]
-        for stat, key in (
-            ('rms', 'rms'),
-            ('mad', 'mad'),
-            ('mean', 'mean'),
-            ('p95_abs', 'p95_abs'),
-        ):
-            stat_value = _residual_stat(values, stat)
-            if stat_value is not None:
-                out[key][cell_id] = float(stat_value)
-    return {
-        key: np.ascontiguousarray(value, dtype=np.float64)
-        for key, value in out.items()
-    }
-
+def _per_cell_residual_stats_ms(*args: Any, **kwargs: Any) -> Any:
+    return getattr(_cell_velocity_artifacts(), '_per_cell_residual_stats_ms')(*args, **kwargs)
 
 def _qc_int(qc: dict[str, Any], key: str, *, default: int) -> int:
     raw = qc.get(key)
@@ -6661,7 +5176,6 @@ def _qc_int(qc: dict[str, Any], key: str, *, default: int) -> int:
         raise RefractionStaticArtifactError(
             f'QC field {key} must be an integer'
         ) from exc
-
 
 def _required_positive_qc_int(qc: dict[str, Any], key: str) -> int:
     raw = qc.get(key)
@@ -6678,7 +5192,6 @@ def _required_positive_qc_int(qc: dict[str, Any], key: str) -> int:
             f'QC field {key} must be a positive integer'
         )
     return value
-
 
 def _qc_cell_id_array(
     qc: dict[str, Any],
@@ -6703,7 +5216,6 @@ def _qc_cell_id_array(
             f'QC field {key} contains out-of-range cell IDs'
         )
     return out
-
 
 def _qc_cell_count_array(
     qc: dict[str, Any],
@@ -6731,7 +5243,6 @@ def _qc_cell_count_array(
         )
     return out
 
-
 def _qc_optional_float(
     qc: dict[str, Any],
     key: str,
@@ -6741,13 +5252,11 @@ def _qc_optional_float(
     raw = qc.get(key, default)
     return _json_float(raw)
 
-
 def _length(value: object, *, name: str) -> int:
     arr = np.asarray(value)
     if arr.ndim != 1:
         raise RefractionStaticArtifactError(f'{name} must be one-dimensional')
     return int(arr.shape[0])
-
 
 def _validate_status_array(value: object, *, name: str) -> None:
     unknown = sorted(
@@ -6762,31 +5271,24 @@ def _validate_status_array(value: object, *, name: str) -> None:
             f'unknown status array values in {name}: {unknown}'
         )
 
-
 def _scalar_str(value: object) -> np.ndarray:
     text = '' if value is None else str(value)
     return np.asarray(text, dtype=f'<U{max(1, len(text))}')
 
-
 def _scalar_int(value: object) -> np.ndarray:
     return np.asarray(int(value), dtype=np.int64)
-
 
 def _scalar_float(value: object) -> np.ndarray:
     return np.asarray(float(value), dtype=np.float64)
 
-
 def _int_array(value: object) -> np.ndarray:
     return np.ascontiguousarray(value, dtype=np.int64)
-
 
 def _float_array(value: object) -> np.ndarray:
     return np.ascontiguousarray(value, dtype=np.float64)
 
-
 def _filled_float_array(value: object, shape: int) -> np.ndarray:
     return np.full(int(shape), float(value), dtype=np.float64)
-
 
 def _endpoint_v2_m_s(
     value: object,
@@ -6798,24 +5300,20 @@ def _endpoint_v2_m_s(
         return _filled_float_array(scalar_v2_m_s, shape)
     return np.ascontiguousarray(value, dtype=np.float64)
 
-
 def _endpoint_cell_id_array(value: object, shape: int) -> np.ndarray:
     if value is None:
         return np.full(int(shape), -1, dtype=np.int64)
     return np.ascontiguousarray(value, dtype=np.int64)
-
 
 def _cell_id_float_array(value: object) -> np.ndarray:
     out = np.asarray(value, dtype=np.float64).copy()
     out[out < 0] = np.nan
     return np.ascontiguousarray(out, dtype=np.float64)
 
-
 def _endpoint_v2_status_array(value: object, shape: int) -> np.ndarray:
     if value is None:
         return _string_array(np.full(int(shape), 'ok', dtype='<U2'))
     return _string_array(value)
-
 
 def _sum_float_arrays(left: object, right: object) -> np.ndarray:
     left_arr = np.asarray(left, dtype=np.float64)
@@ -6825,16 +5323,13 @@ def _sum_float_arrays(left: object, right: object) -> np.ndarray:
     out[finite] = left_arr[finite] + right_arr[finite]
     return np.ascontiguousarray(out, dtype=np.float64)
 
-
 def _bool_array(value: object) -> np.ndarray:
     return np.ascontiguousarray(value, dtype=bool)
-
 
 def _string_array(value: object) -> np.ndarray:
     raw = [str(item) for item in np.asarray(value).tolist()]
     max_len = max([1, *(len(item) for item in raw)])
     return np.ascontiguousarray(raw, dtype=f'<U{max_len}')
-
 
 def _sum_correction_s(left: object, right: object) -> float:
     left_value = _float_or_nan(left)
@@ -6842,7 +5337,6 @@ def _sum_correction_s(left: object, right: object) -> float:
     if not np.isfinite(left_value) or not np.isfinite(right_value):
         return float('nan')
     return float(left_value + right_value)
-
 
 from app.services.refraction_static_artifacts.components import (  # noqa: E402
     write_refraction_static_component_qc_artifacts,  # noqa: F811
@@ -6877,7 +5371,6 @@ from app.services.refraction_static_artifacts.line_profile import (  # noqa: E40
     build_refraction_line_profile_qc_payload as build_refraction_line_profile_qc_payload,
     write_refraction_line_profile_qc_artifacts as write_refraction_line_profile_qc_artifacts,
 )
-
 
 __all__ = [
     'FIRST_BREAK_RESIDUALS_CSV_NAME',
