+++ app/tests/test_refraction_static_artifacts.py
@@ -12,6 +12,7 @@ import pytest
 from app.api.schemas import RefractionStaticApplyRequest
 import app.services.refraction_static_artifacts as artifact_module
 import app.services.refraction_static_artifacts._legacy as artifact_legacy_module
+import app.services.refraction_static_artifacts.cell_velocity as cell_velocity_module
 from app.services.refraction_static_artifacts import registry as artifact_registry
 from app.services.refraction_static_artifacts.io import (
     _write_csv_atomic,
@@ -406,6 +407,20 @@ def test_refraction_static_artifacts_public_all_snapshot() -> None:
         'write_source_static_table_csv',
     )
     assert all(hasattr(artifact_module, name) for name in artifact_module.__all__)
+
+
+def test_cell_velocity_artifact_entry_points_are_in_cell_velocity_module() -> None:
+    expected = (
+        'build_refraction_cell_solver_history_rows',
+        'build_refraction_refractor_velocity_grid_arrays',
+        'build_refraction_refractor_velocity_qc_payload',
+        'write_refraction_cell_solver_history_csv',
+        'write_refraction_refractor_velocity_cells_csv',
+        'write_refraction_refractor_velocity_grid_npz',
+        'write_refraction_refractor_velocity_qc_json',
+    )
+    assert all(hasattr(cell_velocity_module, name) for name in expected)
+    assert all(getattr(artifact_module, name).__module__.endswith('.cell_velocity') for name in expected)
 
 
 def test_refraction_static_artifacts_contract_constants_are_direct_importable() -> None:
